Calculate Average Marginal Effects (AMEs) from brms models

This function is designed to help calculate average marginal effects (AMEs) from brms models. Currently, only one of at or add can be specified. It is hoped to relax this in the future. TODO add more documentation and technical definitions.

brmsmargins(
  object,
  at = NULL,
  add = NULL,
  newdata = model.frame(object),
  CI = 0.99,
  CIType = "HDI",
  contrasts = NULL,
  ROPE = NULL,
  MID = NULL,
  subset = NULL,
  ...
)

Arguments

object
A fitted brms model object. Required.

at
An optional object inheriting from data frame indicating the values to hold specific variables at when calculating average predictions. This is intended for AMEs from categorical variables.

add
An optional object inheriting from data frame indicating the values to add to specific variables at when calculating average predictions. This is intended for AMEs for continuous variables.

newdata
An object inheriting from data frame indicating the baseline values to use for predictions and AMEs. Defaults to be the model frame.

CI
A numeric value specifying the width of the credible interval. Defaults to 0.99.

CIType
A character string specifying the type of credible interval (e.g., highest density interval). It is passed down to bsummary which in turn passes it to ci. Defaults to “HDI”.

contrasts
An optional contrast matrix. The posterior predictions matrix is post multiplied by the contrast matrix, so they must be conformable. The posterior predictions matrix has a separate column for each row in the at or add object, so the contrast matrix should have the same number of rows. It can have multiple columns, if you desire multiple specific contrasts.

ROPE
Either left as NULL, the default, or a numeric vector of length 2, specifying the lower and upper thresholds for the Region of Practical Equivalence (ROPE).

MID
Either left as NULL, the default, or a numeric vector of length 2, specifying the lower and upper thresholds for a Minimally Important Difference (MID). Unlike the ROPE, percentages for the MID are calculated as at or exceeding the bounds specified by this argument, whereas the ROPE is the percentage of the posterior at or inside the bounds specified.

subset
A character string that is a valid R expression used to subset the dataset passed in newdata, prior to analysis. Defaults to NULL.

...
Additional arguments passed on to .predict.

Value

A list. TODO describe more.

Examples

# \dontrun{
#### Testing ####
## sample data and logistic model with brms
set.seed(1234)
Tx <- rep(0:1, each = 50)
ybin <- c(rep(0:1, c(40,10)), rep(0:1, c(10,40)))
logitd <- data.frame(Tx = Tx, ybin = ybin)
logitd$x <- rnorm(100, mean = logitd$ybin, sd = 2)

mbin <- brms::brm(ybin ~ Tx + x, data = logitd, family = brms::bernoulli())
#> Compiling Stan program...
#> Start sampling
#> 
#> SAMPLING FOR MODEL 'fd9b113038d6c3d015532b47d187dbe5' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.037 seconds (Warm-up)
#> Chain 1:                0.039 seconds (Sampling)
#> Chain 1:                0.076 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'fd9b113038d6c3d015532b47d187dbe5' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.037 seconds (Warm-up)
#> Chain 2:                0.031 seconds (Sampling)
#> Chain 2:                0.068 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'fd9b113038d6c3d015532b47d187dbe5' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.035 seconds (Warm-up)
#> Chain 3:                0.039 seconds (Sampling)
#> Chain 3:                0.074 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'fd9b113038d6c3d015532b47d187dbe5' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.036 seconds (Warm-up)
#> Chain 4:                0.031 seconds (Sampling)
#> Chain 4:                0.067 seconds (Total)
#> Chain 4: 

summary(mbin)
#>  Family: bernoulli 
#>   Links: mu = logit 
#> Formula: ybin ~ Tx + x 
#>    Data: logitd (Number of observations: 100) 
#>   Draws: 4 chains, each with iter = 2000; warmup = 1000; thin = 1;
#>          total post-warmup draws = 4000
#> 
#> Population-Level Effects: 
#>           Estimate Est.Error l-95% CI u-95% CI Rhat Bulk_ESS Tail_ESS
#> Intercept    -1.28      0.38    -2.09    -0.58 1.00     4352     2969
#> Tx            2.51      0.54     1.48     3.60 1.00     4105     2865
#> x             0.38      0.14     0.11     0.67 1.00     3580     3061
#> 
#> Draws were sampled using sampling(NUTS). For each parameter, Bulk_ESS
#> and Tail_ESS are effective sample size measures, and Rhat is the potential
#> scale reduction factor on split chains (at convergence, Rhat = 1).

## predictions + summary 
test1 <- brmsmargins:::.predict(mbin,
         model.frame(mbin),
         summarize = TRUE, posterior = FALSE, dpar = NULL, re_formula = NULL,
         resample = 0L)
test1
#> $Summary
#>            M      Mdn        LL        UL PercentROPE PercentMID   CI CIType
#> 1: 0.5002456 0.500962 0.4049813 0.6023425          NA         NA 0.99    HDI
#>    ROPE  MID
#> 1: <NA> <NA>
#> 

## check that bootstrapping the sample / population assumed as part of the
## AME indeed increases the uncertainty interval
## TODO: point estimates (M, Mdn) should be based on the actual data
## not the bootstrapped
test2 <- brmsmargins:::.predict(mbin,
         model.frame(mbin),
         summarize = TRUE, posterior = FALSE, dpar = NULL, re_formula = NULL,
         resample = 100L, seed = 1234)
test2
#> $Summary
#>            M       Mdn        LL        UL PercentROPE PercentMID   CI CIType
#> 1: 0.5001863 0.5005524 0.3716074 0.6278819          NA         NA 0.99    HDI
#>    ROPE  MID
#> 1: <NA> <NA>
#> 

## now check AME for Tx
tmp <- brmsmargins(
  object = mbin,
  at = data.table::data.table(Tx = 0:1),
  contrasts = matrix(c(-1, 1), nrow = 2),
  ROPE = c(-.05, +.05),
  MID = c(-.10, +.10))

tmp$Summary
#>            M       Mdn        LL        UL PercentROPE PercentMID   CI CIType
#> 1: 0.2608257 0.2597441 0.1130868 0.4352608           0      99.65 0.99    HDI
#> 2: 0.7491560 0.7523842 0.5769914 0.8966183           0     100.00 0.99    HDI
#>             ROPE                       MID
#> 1: [-0.05, 0.05] [-Inf, -0.1] | [0.1, Inf]
#> 2: [-0.05, 0.05] [-Inf, -0.1] | [0.1, Inf]
tmp$ContrastSummary ## Tx AME
#> [[1]]
#>            M       Mdn        LL        UL PercentROPE PercentMID   CI CIType
#> 1: 0.4883303 0.4884615 0.2436184 0.7196581           0        100 0.99    HDI
#>             ROPE                       MID
#> 1: [-0.05, 0.05] [-Inf, -0.1] | [0.1, Inf]
#> 

## now check AME for Tx with bootstrapping the AME population
tmpalt <- brmsmargins(
  object = mbin,
  at = data.table::data.table(Tx = 0:1),
  contrasts = matrix(c(-1, 1), nrow = 2),
  ROPE = c(-.05, +.05),
  MID = c(-.10, +.10),
  resample = 100L)

tmpalt$Summary
#>            M       Mdn        LL        UL PercentROPE PercentMID   CI CIType
#> 1: 0.2607922 0.2585965 0.1017346 0.4346288           0     99.645 0.99    HDI
#> 2: 0.7491551 0.7529014 0.5710419 0.8969755           0    100.000 0.99    HDI
#>             ROPE                       MID
#> 1: [-0.05, 0.05] [-Inf, -0.1] | [0.1, Inf]
#> 2: [-0.05, 0.05] [-Inf, -0.1] | [0.1, Inf]
tmpalt$ContrastSummary ## Tx AME
#> [[1]]
#>            M       Mdn        LL        UL PercentROPE PercentMID   CI CIType
#> 1: 0.4883628 0.4901685 0.2332385 0.7235288           0    99.9905 0.99    HDI
#>             ROPE                       MID
#> 1: [-0.05, 0.05] [-Inf, -0.1] | [0.1, Inf]
#> 

## now check AME for continuous predictor, x
## use .01 as an approximation for first derivative
## 1 / .01 in the contrast matrix to get back to a one unit change metric
tmp2 <- brmsmargins(
  object = mbin,
  add = data.table::data.table(x = c(0, .01)),
  contrasts = matrix(c(-1/.01, 1/.01), nrow = 2),
  ROPE = c(-.05, +.05),
  MID = c(-.10, +.10))

tmp2$ContrastSummary ## x AME
#> [[1]]
#>             M        Mdn          LL         UL PercentROPE PercentMID   CI
#> 1: 0.05310642 0.05348044 0.005255614 0.09772485      42.675        0.4 0.99
#>    CIType          ROPE                       MID
#> 1:    HDI [-0.05, 0.05] [-Inf, -0.1] | [0.1, Inf]
#> 

if (FALSE) {
  library(lme4)
  data(sleepstudy)
  fit <- brms::brm(Reaction ~ 1 + Days + (1+ Days | Subject), 
             data = sleepstudy,
             cores = 4)

  summary(fit) 

  tmp <- brmsmargins(
    object = fit,
    at = data.table::data.table(Days = 0:1),
    contrasts = matrix(c(-1, 1), nrow = 2),
    ROPE = c(-.05, +.05),
    MID = c(-.10, +.10))

  tmp$Summary
  tmp$ContrastSummary
  }
# }