Convert a Link Function Name to a List

Internal utility function used in prediction(). Takes a link function name as a character string, the type of effect to be used, and the desired back transformation and returns a list with all the options needed to execute the desired options in prediction().

Usage

.links(
  link,
  effects = c("fixedonly", "includeRE", "integrateoutRE"),
  backtrans = c("response", "linear", "identity", "invlogit", "exp", "square", "inverse")
)

Arguments

link

The link named in a brms::brmsfit object

effects

A character string, the type of effect desired

backtrans

A character string, the type of back transformation

Value

A list with eight elements.

scale A character string giving the argument to be passed to fitted().

ilink A character string giving the name of the inverse link function.

ifun Inverse link function as an R function.

ilinknum An integer giving the inverse link / transformation to be applied in integratere(), needed as this is a C++ function and cannot use the R based inverse link function.