Function to generate posterior predictions and summaries from a brms model with random effects

This is an early developmental function.

.predictre(
  object,
  data,
  backtrans = NULL,
  index,
  summarize = TRUE,
  posterior = FALSE,
  dpar = NULL,
  re_formula = NULL,
  k = 100L,
  resample = 0L,
  seed,
  ...
)

Arguments

object
A fitted brms model object. Required.

data
A data frame or data table passed to fitted() as the new data to be used for predictions. Required.

summarize
A logical value, whether or not to calculate summaries of the posterior predictions. Defaults to TRUE.

posterior
A logical value whether or not to save and return the posterior samples. Defaults to FALSE as the assumption is a typical use case is to return the summaries only.

dpar
Parameter passed on the dpar argument of fitted() in brms.

re_formula
Parameter passed on the re_formula argument of fitted() in brms.

resample
An integer indicating the number of bootstrap resamples of the posterior predictions to use when calculating summaries. Defaults to 0L. See the details section for more informations as its implementation is experimental and it may not operate as one would expect.

seed
A seed for random number generation. Missing by default. Only needed if resample is a non zero integer.

...
Additional arguments passed to fitted()

Value

A list with Summary and Posterior. Some of these may be NULL depending on the arguments used.