Integrate over multivariate normal random effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>           [,1]     [,2]     [,3]      [,4]     [,5]       [,6]      [,7]
#> [1,] -4.274115 14.62454 6.920974 -20.49348 2.043359 -0.5658749 0.7076795
#>           [,8]     [,9]     [,10]    [,11]     [,12]    [,13]    [,14]    [,15]
#> [1,] -39.14412 1.121047 -14.33089 2.190016 -4.119444 12.65164 2.765349 23.04636
#>         [,16]     [,17]      [,18]    [,19]    [,20]     [,21]    [,22]
#> [1,] -6.27379 -5.180588 -0.3955865 21.44424 18.34887 -12.75065 3.647414
#>         [,23]    [,24]   [,25]    [,26]     [,27]    [,28]    [,29]    [,30]
#> [1,] 3.260383 -1.21229 13.0119 1.822789 -10.14384 22.74473 5.177685 8.703213
#>           [,31]     [,32]     [,33]     [,34]     [,35]      [,36]    [,37]
#> [1,] -0.5794792 0.9183164 -22.54197 -21.40043 -15.75907 -0.2759706 28.13097
#>         [,38]    [,39]    [,40]     [,41]     [,42]     [,43]     [,44]
#> [1,] 6.489213 1.318117 2.266877 -17.33354 -14.75618 -12.14711 -23.37348
#>          [,45]     [,46]     [,47]     [,48]    [,49]     [,50]    [,51]
#> [1,] 0.3646962 0.4462169 -5.741837 -7.860372 28.31668 -2.834677 -10.7933
#>           [,52]    [,53]    [,54]    [,55]     [,56]    [,57]    [,58]   [,59]
#> [1,] -0.9017812 1.270546 9.082999 9.441627 -6.932917 32.29667 18.94786 11.4781
#>          [,60]    [,61]     [,62]    [,63]    [,64]    [,65]       [,66]
#> [1,] -25.87971 5.511757 -25.99933 14.17742 18.51695 17.75356 0.001989756
#>         [,67]    [,68]   [,69]    [,70]    [,71]    [,72]     [,73]    [,74]
#> [1,] 19.62093 19.91456 8.93752 6.681913 11.41615 11.24231 -9.701548 11.57137
#>          [,75]   [,76]     [,77]     [,78]     [,79]      [,80]    [,81]
#> [1,] -5.212463 2.26606 -6.415474 -10.29601 -22.69166 -0.1822295 9.546411
#>         [,82]    [,83]    [,84]     [,85]     [,86]    [,87]     [,88]
#> [1,] 19.44706 6.227569 6.425704 -1.121849 -22.13059 9.962674 -7.587288
#>          [,89]    [,90]     [,91]     [,92]    [,93]     [,94]     [,95]
#> [1,] -11.61402 19.20243 -5.412847 0.7219623 2.046283 -28.32703 -11.62624
#>         [,96]    [,97]    [,98]     [,99]   [,100]
#> [1,] 21.91413 13.94033 11.41608 -7.354611 6.385032

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>          [,1]      [,2]      [,3]     [,4]     [,5]     [,6]      [,7]
#> [1,] 7.402808 -1.207278 -7.473032 1.620572 0.997979 2.039059 -4.193989
#>           [,8]      [,9]    [,10]   [,11]   [,12]    [,13]    [,14]      [,15]
#> [1,] -1.264036 -10.07635 2.297408 3.27199 7.14866 2.109975 6.639945 -0.7377623
#>          [,16]     [,17]     [,18]     [,19]     [,20]    [,21]    [,22]
#> [1,] 0.9789128 -12.55141 -6.388532 -8.041471 0.4733152 2.138638 1.620685
#>          [,23]    [,24]     [,25]     [,26]    [,27]    [,28]     [,29]
#> [1,] -5.502372 9.253957 -2.402123 -9.674564 4.835841 -4.28189 -4.763979
#>          [,30]    [,31]    [,32]     [,33]     [,34]     [,35]    [,36]
#> [1,] -2.275018 8.816575 2.101295 -14.03877 -2.386007 -7.274981 2.014947
#>         [,37]     [,38]   [,39]     [,40]    [,41]    [,42]     [,43]    [,44]
#> [1,] 1.160463 -3.073256 5.05296 -1.871148 1.724055 1.849114 -4.775231 8.388598
#>          [,45]     [,46]    [,47]    [,48]    [,49]     [,50]     [,51]   [,52]
#> [1,] -1.777878 -4.113012 3.821487 3.380403 6.742841 -5.416654 -3.794255 4.04035
#>          [,53]    [,54]      [,55]     [,56]    [,57]    [,58]    [,59]
#> [1,] -3.449988 7.212656 -0.8311062 -2.472808 1.216926 8.158667 5.515159
#>         [,60]    [,61]    [,62]     [,63]    [,64]    [,65]    [,66]    [,67]
#> [1,] 2.897372 6.130439 2.538557 -3.759073 1.344105 4.755553 6.011672 -8.84544
#>          [,68]    [,69]    [,70]      [,71]     [,72]    [,73]    [,74]
#> [1,] -5.182652 11.69091 -7.89123 -0.6390108 -2.027809 1.668979 11.71408
#>          [,75]    [,76]    [,77]     [,78]     [,79]    [,80]     [,81]
#> [1,] -4.017113 9.138745 3.954578 -1.930172 -11.08578 8.503743 -1.060823
#>         [,82]     [,83]    [,84]     [,85]     [,86]    [,87]     [,88]
#> [1,] -2.37102 -4.786336 12.11193 -3.764694 -2.562838 5.800204 -7.529181
#>          [,89]     [,90]      [,91]     [,92]    [,93]      [,94]     [,95]
#> [1,] -5.916754 0.7651727 -0.6261357 -3.097112 4.816706 -0.7460322 -1.020154
#>          [,96]     [,97]     [,98]     [,99]   [,100]
#> [1,] -5.812364 -1.105832 -1.024603 0.3144244 1.116689