Convert a link function name to a list

Internal utility function used in .predict(). Takes a link function name as a character string, the type of effect to be used, and the desired back transformation and returns a list with all the options needed to execute the desired options in .predict().

Usage

.links(
  link,
  effects = c("fixedonly", "includeRE", "integrateoutRE"),
  backtrans = c("response", "linear", "identity", "invlogit", "exp", "square")
)

Arguments

link

The link named in a brmsfit object

effects

A character string, the type of effect desired

backtrans

A character string, the type of back transformation

Value

A list with four elements.

scale

A character string giving the argument to be passed to fitted().

ilink

A character string giving the name of the inverse link function.

ifun

Inverse link function as an R function.

ilinknum

An integer giving the inverse link / transformation to be applied in integratere(), needed as this is a C++ function and cannot use the R based inverse link function.