Function to generate posterior predictions and summaries from a brms model

This is an internal function that is essentially a fancy wrapper for fitted().

.predict(
  object,
  data,
  summarize = TRUE,
  posterior = FALSE,
  index,
  dpar = NULL,
  resample = 0L,
  seed = FALSE,
  effects = c("fixedonly", "includeRE", "integrateoutRE"),
  backtrans = c("response", "linear", "identity", "invlogit", "exp", "square"),
  k = 100L,
  ...
)

Arguments

object
A fitted brms model object. Required.

data
A data frame or data table passed to fitted() as the new data to be used for predictions. Required.

summarize
A logical value, whether or not to calculate summaries of the posterior predictions. Defaults to TRUE.

posterior
A logical value whether or not to save and return the posterior samples. Defaults to FALSE as the assumption is a typical use case is to return the summaries only.

index
An optional integer vector, giving the posterior draws to be used in the calculations. If omitted, defaults to all posterior draws.

dpar
Parameter passed on the dpar argument of fitted() in brms. Defaults to NULL indicating the mean or location parameter typically.

resample
An integer indicating the number of bootstrap resamples of the posterior predictions to use when calculating summaries. Defaults to 0L. See documentation from .averagePosterior() for more details.

seed
A seed for random number generation. Defaults to FALSE, which means no seed is set. Only used if resample is a positive, non-zero integer. See documentation from .averagePosterior() for more details.

effects
A character string indicating the type of prediction to be made. Can be one of “fixedonly” meaning only use fixed effects, “includeRE” meaning that random effects should be included in the predictions, or “integrateoutRE” meaning that random effects should be integrated out / over in the predictions.

backtrans
A character string indicating the type of back transformation to be applied. Can be one of “response” meaning to use the response scale, “linear” or “identity” meaning to use the linear predictor scale, or a specific back transformation desired, from a possible list of “invlogit”, “exp”, or “square”. Custom back transformations should only be needed if, for example, the outcome variable was transformed prior to fitting the model.

k
An integer providing the number of random draws to use for integrating out the random effects. Only relevant when effects is “integrateoutRE”.

...
Additional arguments passed to fitted()

Value

A list with Summary and Posterior. Some of these may be NULL depending on the arguments used.