Integrate over Multivariate Normal Random Effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>           [,1]     [,2]     [,3]     [,4]      [,5]     [,6]      [,7]
#> [1,] -11.32393 8.222921 17.48488 3.895515 -5.061825 12.51682 0.4564491
#>           [,8]      [,9]    [,10]     [,11]    [,12]   [,13]   [,14]    [,15]
#> [1,] -1.485455 -9.024374 16.55363 -34.50475 15.03721 15.1896 18.8853 -7.51807
#>         [,16]    [,17]     [,18]    [,19]     [,20]     [,21]    [,22]    [,23]
#> [1,] 9.182097 -6.61996 -11.86457 7.679975 -28.16643 -17.48763 6.236124 1.605881
#>          [,24]    [,25]     [,26]     [,27]     [,28]    [,29]     [,30]
#> [1,] 0.3979416 11.30126 0.5948682 -2.984116 -6.215574 31.99465 -11.04466
#>        [,31]     [,32]    [,33]    [,34]     [,35]     [,36]     [,37]
#> [1,] 13.7193 -10.99514 3.478001 5.505992 -20.87441 -13.50353 -8.754594
#>          [,38]    [,39]    [,40]     [,41]    [,42]     [,43]     [,44]
#> [1,] -5.068691 6.249611 23.83101 -12.10276 7.728667 -1.687624 -17.91734
#>          [,45]      [,46]     [,47]    [,48]      [,49]      [,50]   [,51]
#> [1,] -41.30215 -0.3735652 -12.13458 18.56828 -0.8523716 -0.0919671 7.43696
#>         [,52]    [,53]     [,54]     [,55]    [,56]    [,57]     [,58]    [,59]
#> [1,] 2.652263 4.286173 -6.167304 -2.340124 16.48808 -24.0728 -13.79975 -13.0409
#>          [,60]    [,61]    [,62]     [,63]    [,64]    [,65]    [,66]    [,67]
#> [1,] -25.45944 12.65988 15.41035 -4.224627 30.91854 -6.43181 15.68179 2.314497
#>         [,68]    [,69]    [,70]     [,71]    [,72]    [,73]    [,74]    [,75]
#> [1,] 25.56357 20.95169 19.29764 -15.93681 3.620216 5.024768 -4.55304 3.694685
#>         [,76]    [,77]     [,78]    [,79]    [,80]     [,81]    [,82]     [,83]
#> [1,] -1.09793 5.270174 -15.61018 5.341948 18.53258 -7.109613 3.611967 -6.528037
#>         [,84]     [,85]   [,86]     [,87]      [,88]    [,89]     [,90]
#> [1,] 13.02273 -13.84014 2.01377 -4.874746 -0.1002473 14.66798 -6.888748
#>         [,91]     [,92]    [,93]     [,94]    [,95]   [,96]     [,97]    [,98]
#> [1,] 6.757162 -4.586944 2.687152 -10.07718 3.649087 13.8336 -19.39861 9.643754
#>          [,99]   [,100]
#> [1,] 0.2474186 1.992288

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>          [,1]      [,2]     [,3]      [,4]    [,5]      [,6]     [,7]     [,8]
#> [1,] 2.363482 -4.653555 7.126341 -5.473015 1.01408 -2.078928 4.932135 7.598167
#>          [,9]     [,10]    [,11]     [,12]    [,13]     [,14]    [,15]
#> [1,] 2.822423 -1.704703 5.258167 -1.332284 3.982691 -3.240187 4.107054
#>          [,16]    [,17]     [,18]    [,19]       [,20]    [,21]      [,22]
#> [1,] -4.280084 1.801532 -4.628782 2.385498 -0.03086002 7.180268 -0.8029381
#>          [,23]    [,24]      [,25]    [,26]    [,27]     [,28]     [,29]
#> [1,] -10.17091 6.672905 -0.7843831 5.137921 5.881928 -1.943712 -2.489592
#>         [,30]     [,31]     [,32]     [,33]     [,34]   [,35]     [,36]
#> [1,] 1.520663 -3.748297 -4.254246 0.2074237 -4.413813 3.51107 -6.937712
#>           [,37]    [,38]     [,39]    [,40]     [,41]    [,42]     [,43]
#> [1,] -0.1936313 3.740288 0.0429324 3.022721 -5.880638 8.031459 -3.207183
#>         [,44]    [,45]    [,46]     [,47]     [,48]     [,49]    [,50]    [,51]
#> [1,] -8.15713 10.01937 1.263558 -2.890034 -6.340571 -3.653567 2.765581 2.436349
#>          [,52]       [,53]    [,54]    [,55]     [,56]   [,57]    [,58]
#> [1,] -3.307853 -0.02544217 -5.30317 5.935465 -3.480433 7.52083 4.331703
#>          [,59]     [,60]    [,61]     [,62]    [,63]     [,64]     [,65]
#> [1,] -3.565308 -6.973094 4.369966 -4.747655 1.009922 -5.748521 -3.450419
#>          [,66]     [,67]    [,68]    [,69]    [,70]    [,71]     [,72]    [,73]
#> [1,] -2.067407 -2.280928 4.983151 1.319128 1.623939 5.764739 -2.908675 4.321274
#>          [,74]     [,75]    [,76]    [,77]    [,78]     [,79]    [,80]    [,81]
#> [1,] -1.216139 -1.566716 1.235663 -1.05598 2.500261 -1.363401 2.673928 2.026322
#>         [,82]   [,83]    [,84]     [,85]     [,86]     [,87]     [,88]   [,89]
#> [1,] 7.533101 1.49449 4.875582 -6.880606 -3.432227 -3.820403 -2.806158 3.89794
#>         [,90]    [,91]     [,92]     [,93]     [,94]    [,95]     [,96]
#> [1,] -1.45718 6.316397 -1.099915 -0.273937 -1.044748 3.481534 0.1638671
#>         [,97]     [,98]     [,99]    [,100]
#> [1,] 13.83602 -2.812351 -2.770373 -5.013133