Integrate over multivariate normal random effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>         [,1]     [,2]      [,3]     [,4]      [,5]      [,6]      [,7]
#> [1,] 8.26875 9.289201 -21.00875 4.026489 -6.802197 -11.55882 -51.49868
#>           [,8]      [,9]     [,10]    [,11]     [,12]    [,13]     [,14]
#> [1,] -2.343133 -3.847263 -5.066584 21.29691 -22.95879 2.117697 -17.72175
#>          [,15]    [,16]    [,17]     [,18]    [,19]    [,20]     [,21]    [,22]
#> [1,] -13.55476 8.231834 -4.58925 -15.37195 14.07454 5.108941 -4.104976 11.16587
#>         [,23]     [,24]      [,25]      [,26]     [,27]    [,28]     [,29]
#> [1,] 18.25192 -6.803455 -0.9229297 -0.2784732 -3.677056 11.70982 -8.448177
#>        [,30]     [,31]    [,32]     [,33]     [,34]   [,35]    [,36]     [,37]
#> [1,] 15.5623 -2.739421 2.133854 -14.10635 -4.454956 21.0585 4.655885 -22.97351
#>          [,38]   [,39]     [,40]     [,41]    [,42]     [,43]      [,44]
#> [1,] 0.5545007 12.4838 -9.702899 -2.716217 6.725245 -41.66349 -0.3205869
#>          [,45]    [,46]    [,47]     [,48]  [,49]    [,50]    [,51]    [,52]
#> [1,] -13.83346 10.55886 16.48131 -4.643885 6.2074 4.555307 9.451652 4.832545
#>         [,53]    [,54]     [,55]    [,56]     [,57]     [,58]     [,59]
#> [1,] 8.582333 6.544661 -10.41964 -21.2145 -33.88278 -7.537375 -11.25579
#>         [,60]   [,61]    [,62]     [,63]     [,64]    [,65]     [,66]    [,67]
#> [1,] 11.46253 3.85323 21.38939 0.5092463 -1.449192 5.319837 -4.895867 9.469334
#>         [,68]     [,69]     [,70]   [,71]    [,72]     [,73]     [,74]    [,75]
#> [1,] 1.940027 -3.368766 -3.896661 18.6851 5.672638 -9.104238 -11.94623 17.95784
#>         [,76]    [,77]    [,78]    [,79]     [,80]    [,81]    [,82]    [,83]
#> [1,] 2.861104 17.82252 13.87953 6.756468 -29.86081 7.352478 3.379427 13.41298
#>         [,84]    [,85]     [,86]    [,87]     [,88]     [,89]     [,90]
#> [1,] 3.024999 -1.35842 -12.77635 3.549821 -19.43454 -1.307922 -17.53039
#>          [,91]     [,92]     [,93]    [,94]     [,95]      [,96]    [,97]
#> [1,] -2.997321 -22.20445 -5.798877 23.56711 -20.08182 -0.3768457 5.989157
#>        [,98]     [,99]  [,100]
#> [1,] 16.4992 -1.136735 3.13975

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>         [,1]      [,2]       [,3]      [,4]      [,5]       [,6]      [,7]
#> [1,] 3.55343 -4.206618 -0.1553824 -3.985062 -2.596826 -0.4602397 -3.010348
#>          [,8]      [,9]     [,10]    [,11]    [,12]     [,13]     [,14]
#> [1,] 0.521579 -6.740164 -2.652433 -2.68239 1.468268 -6.001097 -3.276102
#>           [,15]    [,16]     [,17]     [,18]     [,19]    [,20]     [,21]
#> [1,] -0.2857142 4.078078 -5.224256 -1.683981 -12.40505 3.350257 -7.152774
#>          [,22]     [,23]    [,24]    [,25]    [,26]       [,27]    [,28]
#> [1,] -6.592816 0.9152813 1.172998 -5.35865 7.039945 -0.03949961 7.783091
#>         [,29]     [,30]     [,31]  [,32]    [,33]    [,34]    [,35]     [,36]
#> [1,] -11.7964 -3.499353 -5.419754 4.1218 7.607549 2.963607 3.635725 -5.446649
#>          [,37]    [,38]     [,39]    [,40]     [,41]    [,42]     [,43]
#> [1,] -2.782624 8.186774 -2.661982 2.362343 -1.491885 3.080731 -1.815465
#>          [,44]     [,45]     [,46]     [,47]    [,48]     [,49]     [,50]
#> [1,] -2.412653 0.1417565 -6.665013 -6.285767 6.610433 -1.848753 -1.585092
#>          [,51]   [,52]    [,53]     [,54]     [,55]    [,56]      [,57]
#> [1,] -8.848153 5.28142 0.430803 -3.038068 -13.18726 4.541879 -0.6956457
#>         [,58]     [,59]     [,60]     [,61]   [,62]     [,63]     [,64]
#> [1,] 0.853232 -6.455048 -5.205147 -1.431335 1.42571 -2.298181 0.2999091
#>           [,65]    [,66]      [,67]    [,68]    [,69]     [,70]      [,71]
#> [1,] -0.4565238 4.828127 -0.1916898 5.567864 -5.78665 -2.674942 -0.5128643
#>          [,72]    [,73]    [,74]     [,75]    [,76]    [,77]    [,78]    [,79]
#> [1,] -3.652078 6.840345 3.026706 -6.459403 3.988764 3.918729 -1.19712 8.940353
#>      [,80]     [,81]    [,82]     [,83]     [,84]     [,85]     [,86]     [,87]
#> [1,] -15.7 -1.936265 1.924124 -3.687758 0.5729853 -5.650378 -2.510134 0.4545256
#>           [,88]    [,89]      [,90]     [,91]      [,92]    [,93]    [,94]
#> [1,] -0.6694651 5.178109 -0.3450321 -3.008989 -0.2229911 -3.38591 -2.32077
#>          [,95]     [,96]     [,97]     [,98]       [,99]    [,100]
#> [1,] 0.5714339 -3.545632 -1.785746 0.8912852 -0.09273643 -9.450984