Integrate over Multivariate Normal Random Effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>           [,1]      [,2]     [,3]      [,4]     [,5]     [,6]      [,7]
#> [1,] -4.987453 -24.70384 7.879209 -15.16236 -15.7096 1.092649 -14.49187
#>           [,8]       [,9]     [,10]     [,11]    [,12]   [,13]    [,14]
#> [1,] -29.30747 -0.9939016 -10.27655 -11.87427 11.75293 15.5359 3.791924
#>         [,15]     [,16]    [,17]     [,18]     [,19]    [,20]     [,21]
#> [1,] 24.48926 -11.36547 27.82422 -13.78079 -12.87159 7.821744 -18.25708
#>         [,22]     [,23]    [,24]    [,25]      [,26]   [,27]     [,28]
#> [1,] 17.52627 -8.576037 13.96753 2.283352 -0.6166188 10.2754 -9.950931
#>          [,29]   [,30]    [,31]    [,32]     [,33]     [,34]     [,35]
#> [1,] -22.19118 -3.2193 11.19524 3.199886 -9.033537 -6.256491 -3.030898
#>          [,36]     [,37]     [,38]    [,39]     [,40]    [,41]    [,42]
#> [1,] -12.59008 -41.72124 -4.863264 10.70766 -5.365851 13.33583 2.061332
#>          [,43]    [,44]    [,45]    [,46]    [,47]      [,48]     [,49]
#> [1,] -8.319545 3.947917 -4.81717 28.78077 26.36013 -0.8102803 -14.92793
#>          [,50]     [,51]     [,52]     [,53]     [,54]    [,55]    [,56]
#> [1,] 0.5475784 -20.55932 -7.034363 -11.98322 -8.378695 14.88141 4.743558
#>          [,57]    [,58]    [,59]    [,60]  [,61]    [,62]    [,63]     [,64]
#> [1,] -7.793115 11.38779 5.547549 1.144538 17.596 4.073371 6.117157 -29.72125
#>          [,65]    [,66]    [,67]    [,68]    [,69]     [,70]    [,71]     [,72]
#> [1,] -22.26421 2.440526 23.52037 5.425301 7.086791 -10.02071 1.978504 -5.979676
#>          [,73]     [,74]    [,75]     [,76]    [,77]   [,78]    [,79]     [,80]
#> [1,] -11.46114 -4.837284 10.00305 -37.81695 -15.1707 20.0697 10.70334 -1.524124
#>           [,81]    [,82]    [,83]    [,84]    [,85]    [,86]    [,87]   [,88]
#> [1,] -0.7968584 11.17089 13.71977 9.630795 -10.3299 8.104661 17.03723 1.41767
#>          [,89]     [,90]    [,91]    [,92]     [,93]     [,94]     [,95]
#> [1,] -1.362561 -2.383115 16.92573 18.24356 -6.985543 -11.03917 -13.83767
#>         [,96]     [,97]    [,98]     [,99]   [,100]
#> [1,] 2.767456 -18.93163 7.446093 -0.092133 15.64184

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>         [,1]       [,2]      [,3]     [,4]      [,5]   [,6]       [,7]     [,8]
#> [1,] 1.13377 -0.1220542 -6.822737 2.746882 0.5208029 2.3785 -0.1538548 10.03226
#>           [,9]    [,10]      [,11]     [,12]   [,13]     [,14]    [,15]
#> [1,] -5.699565 6.900074 -0.4832468 -2.390763 4.91343 -7.560098 1.759175
#>          [,16]      [,17]    [,18]     [,19]     [,20]    [,21]    [,22]
#> [1,] -7.507952 -0.4867631 6.614594 0.8804442 -1.522555 2.891959 4.504953
#>          [,23]     [,24]     [,25]     [,26]     [,27]    [,28]     [,29]
#> [1,] -5.296494 -4.783512 0.8016953 -4.820197 -6.268106 -5.55994 0.9856321
#>         [,30]       [,31]     [,32]     [,33]     [,34]      [,35]     [,36]
#> [1,] 4.999372 -0.08596601 -10.43884 -2.978486 -5.792852 -0.5625005 -6.790575
#>          [,37]    [,38]     [,39]    [,40]     [,41]     [,42]    [,43]
#> [1,] -2.406853 1.200064 -2.646023 7.798952 -2.708411 -3.415023 11.11325
#>         [,44]   [,45]     [,46]    [,47]    [,48]   [,49]     [,50]    [,51]
#> [1,] 1.661052 6.81487 -6.113546 3.826482 1.833571 2.96138 -1.175292 1.245505
#>         [,52]    [,53]    [,54]     [,55]      [,56]     [,57]     [,58]
#> [1,] 2.978403 1.999775 7.275467 -3.615556 -0.1172784 -6.173656 -3.695145
#>         [,59]     [,60]    [,61]     [,62]     [,63]     [,64]     [,65]
#> [1,] -3.47247 -1.788455 6.655002 -9.257893 -5.101112 0.4896469 -1.310986
#>         [,66]     [,67]     [,68]    [,69]    [,70]     [,71]    [,72]
#> [1,] 1.672825 -0.337617 -2.025775 7.867544 5.659875 -4.660922 4.665071
#>           [,73]     [,74]      [,75]      [,76]     [,77]    [,78]   [,79]
#> [1,] -0.6883455 -4.666849 0.07899648 -0.9918967 -5.467606 2.796579 4.46493
#>           [,80]     [,81]     [,82]    [,83]      [,84]     [,85]     [,86]
#> [1,] -0.7335655 -1.838082 -1.253417 7.307181 -0.7395625 -4.334407 -2.997787
#>          [,87]    [,88]     [,89]     [,90]    [,91]     [,92]    [,93]
#> [1,] -3.712831 8.439616 0.8409801 -3.198061 2.096018 -3.656204 1.868333
#>          [,94]    [,95]    [,96]    [,97]    [,98]     [,99]    [,100]
#> [1,] -4.607901 6.183371 3.331465 12.25427 5.648253 0.4719158 -2.741217