Integrate over multivariate normal random effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>             [,1]     [,2]      [,3]      [,4]     [,5]      [,6]     [,7]
#> [1,] -0.02171253 18.38829 -1.816109 -16.98089 13.10807 -8.939576 2.552223
#>          [,8]      [,9]    [,10]     [,11]    [,12]     [,13]     [,14]
#> [1,] 5.135231 -31.13002 7.494876 -1.435477 7.141254 -14.22501 -12.04998
#>         [,15]     [,16]    [,17]    [,18]     [,19]    [,20]    [,21]    [,22]
#> [1,] 2.029638 -2.505092 13.70723 2.050463 -1.085789 9.387692 7.967533 18.68658
#>        [,23]    [,24]     [,25]     [,26]    [,27]     [,28]     [,29]    [,30]
#> [1,] 17.4932 22.47828 -40.10513 -5.920005 35.14466 -1.538448 -9.954209 29.36084
#>          [,31]     [,32]     [,33]    [,34]   [,35]     [,36]     [,37]
#> [1,] -24.93312 -1.270547 -12.37719 7.667498 6.69531 -23.90968 -27.42367
#>         [,38]    [,39]     [,40]     [,41]    [,42]    [,43]    [,44]     [,45]
#> [1,] 11.17346 21.28047 -24.29359 -9.634553 27.45573 16.71986 9.076992 -6.739417
#>          [,46]     [,47]   [,48]     [,49]    [,50]     [,51]     [,52]
#> [1,] -9.582365 -15.73597 5.26682 0.1358322 5.610018 -25.66219 -8.358068
#>         [,53]   [,54]    [,55]    [,56]    [,57]    [,58]    [,59]     [,60]
#> [1,] -4.51125 12.9094 16.88652 24.46715 10.61677 26.74774 19.77976 -19.21409
#>          [,61]  [,62]     [,63]    [,64]     [,65]    [,66]     [,67]     [,68]
#> [1,] -5.648115 1.4336 0.8969417 2.796221 -1.568987 4.124104 -8.956026 -3.236679
#>          [,69]    [,70]     [,71]     [,72]   [,73]     [,74]     [,75]
#> [1,] -21.93496 16.82368 -2.458639 -25.81423 -8.8237 -13.71584 -20.18928
#>          [,76]     [,77]    [,78]     [,79]    [,80]    [,81]     [,82]
#> [1,] -13.82465 -5.442172 10.61615 -12.49289 9.031805 3.912385 -1.297631
#>         [,83]    [,84]      [,85]     [,86]    [,87]     [,88]     [,89]
#> [1,] 3.926733 23.00207 -0.7308521 -1.541435 25.09013 -9.215211 0.1046494
#>          [,90]    [,91]    [,92]    [,93]     [,94]    [,95]    [,96]    [,97]
#> [1,] -11.15857 5.186598 6.929802 12.62358 -5.680557 11.50163 -4.69964 10.05924
#>         [,98]     [,99]  [,100]
#> [1,] 26.61745 -14.67541 17.0249

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>          [,1]       [,2]      [,3]       [,4]     [,5]     [,6]      [,7]
#> [1,] 7.838487 0.07330064 0.6151499 -0.6879608 3.891877 3.371293 -3.824296
#>           [,8]     [,9]    [,10]    [,11]     [,12]    [,13]    [,14]    [,15]
#> [1,] -6.965183 4.256394 2.115121 5.308887 -1.241532 2.398973 4.318974 4.674837
#>         [,16]     [,17]    [,18]     [,19]   [,20]     [,21]    [,22]     [,23]
#> [1,] 6.082439 -2.137147 1.571101 0.6494289 6.17145 -5.810277 5.335656 -6.374373
#>         [,24]     [,25]    [,26]     [,27]    [,28]    [,29]   [,30]      [,31]
#> [1,] -2.80552 0.3969038 2.653629 -6.866527 6.944365 1.527324 1.97764 -0.0296371
#>         [,32]     [,33]    [,34]    [,35]     [,36]    [,37]     [,38]
#> [1,] 0.190468 -4.658716 2.477594 2.840632 -6.119472 3.033738 -3.202441
#>          [,39]     [,40]      [,41]   [,42]    [,43]     [,44]      [,45]
#> [1,] -7.280033 0.4064649 -0.2918202 4.93291 5.487421 -6.115578 -0.4862717
#>         [,46]    [,47]     [,48]     [,49]    [,50]   [,51]     [,52]     [,53]
#> [1,] 3.655281 1.719192 -2.192078 -4.761899 3.307612 9.58823 -9.513171 -6.901337
#>          [,54]     [,55]     [,56]    [,57]    [,58]    [,59]      [,60]
#> [1,] -5.661263 -1.469566 -1.122049 9.555837 6.857575 2.831366 -0.5183201
#>         [,61]    [,62]    [,63]     [,64]     [,65]     [,66]     [,67]
#> [1,] 2.846418 6.653633 8.133406 0.5280284 -6.999445 0.1560957 -2.121452
#>          [,68]    [,69]    [,70]    [,71]    [,72]     [,73]      [,74]
#> [1,] -2.403246 3.099436 7.725571 1.581908 3.955385 -2.885915 -0.2069842
#>         [,75]    [,76]      [,77]    [,78]     [,79]     [,80]    [,81]
#> [1,] 1.471455 3.175993 -0.2766737 5.895069 -0.457835 -1.280579 4.799372
#>         [,82]     [,83]     [,84]     [,85]     [,86]    [,87]    [,88]
#> [1,] 5.832071 -4.330677 -1.056919 -3.260809 -4.137498 2.415482 3.890966
#>           [,89]    [,90]    [,91]    [,92]     [,93]     [,94]     [,95]
#> [1,] -0.9642686 4.202268 6.217824 9.767858 -4.406186 0.8703053 -5.965985
#>          [,96]     [,97]    [,98]     [,99]   [,100]
#> [1,] -2.671659 -8.726996 -8.22205 -1.167951 -2.31172