Integrate over Multivariate Normal Random Effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>          [,1]      [,2]     [,3]      [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 31.54267 -14.41237 15.76717 -10.06056 1.390265 5.625381 -7.92487 -7.14906
#>            [,9]    [,10]     [,11]     [,12]     [,13]    [,14]     [,15]
#> [1,] -0.0999985 0.530369 -23.26147 -15.99179 -6.364276 -26.3961 -8.831427
#>        [,16]    [,17]     [,18]    [,19]     [,20]    [,21]     [,22]     [,23]
#> [1,] 5.77663 7.938019 0.6151762 10.70873 -20.76006 15.41719 -13.31271 -14.81085
#>         [,24]     [,25]     [,26]      [,27]     [,28]    [,29]    [,30]
#> [1,] 2.936468 0.4804424 -18.96001 -0.5017635 -6.506511 3.363255 4.949462
#>          [,31]     [,32]     [,33]     [,34]     [,35]    [,36]    [,37]
#> [1,] -7.439571 -8.681801 -15.46649 0.3263881 -32.30792 5.851048 9.531791
#>          [,38]     [,39]     [,40]    [,41]    [,42]     [,43]     [,44]
#> [1,] -7.438924 -3.760502 -21.88219 12.27317 13.25018 -15.42045 -26.20419
#>         [,45]     [,46]     [,47]    [,48]     [,49]     [,50]    [,51]
#> [1,] 8.645368 -8.219406 -6.404428 7.760002 -4.369514 -3.053191 8.278134
#>         [,52]    [,53]    [,54]  [,55]    [,56]     [,57]     [,58]    [,59]
#> [1,] 8.089081 2.680427 15.88808 5.4422 2.431559 -25.18703 -7.724366 27.88663
#>          [,60]      [,61]   [,62]    [,63]     [,64]    [,65]     [,66]
#> [1,] -12.42107 -0.7075703 17.5987 10.29771 -16.23618 10.82086 -24.30161
#>          [,67]   [,68]    [,69]    [,70]    [,71]    [,72]     [,73]     [,74]
#> [1,] -22.04986 34.4339 14.25939 13.12147 16.93061 25.80891 -3.920996 -3.691834
#>         [,75]     [,76]     [,77]    [,78]   [,79]    [,80]     [,81]     [,82]
#> [1,] 8.131784 -3.073766 -19.38844 3.252034 37.5031 5.471756 -17.51779 -2.234508
#>          [,83]    [,84]     [,85]     [,86]    [,87]    [,88]    [,89]    [,90]
#> [1,] 0.0764542 1.854554 -10.64204 -1.827287 14.27362 8.056295 2.029139 31.46552
#>           [,91]    [,92]    [,93]     [,94]     [,95]     [,96]    [,97]
#> [1,] -0.9880686 3.176857 8.918931 0.5089416 -9.105024 -21.74045 38.44141
#>         [,98]     [,99]   [,100]
#> [1,] 17.57659 -3.536061 13.69363

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>           [,1]      [,2]      [,3]     [,4]      [,5]      [,6]       [,7]
#> [1,] -2.667193 0.8342411 -4.102745 4.493301 -1.792185 -2.456075 -0.2116305
#>           [,8]       [,9]     [,10]     [,11]    [,12]     [,13]      [,14]
#> [1,] -1.511088 -0.8164794 -2.131601 0.8519648 5.985017 -8.984552 -0.4478739
#>          [,15]      [,16]     [,17]    [,18]     [,19]     [,20]    [,21]
#> [1,] 0.2274685 -0.5573342 -1.045325 5.908076 -3.789153 -4.964417 2.305209
#>         [,22]     [,23]     [,24]     [,25]    [,26]    [,27]     [,28]
#> [1,] -4.16245 -2.857703 -4.721881 -8.492556 1.503172 5.894804 -1.455104
#>           [,29]     [,30]    [,31]      [,32]    [,33]    [,34]    [,35]
#> [1,] -0.6843191 -2.936284 2.402422 -0.3069937 11.34311 3.616981 11.43831
#>          [,36]      [,37]     [,38]     [,39]     [,40]    [,41]     [,42]
#> [1,] -1.747311 -0.1035872 -7.093003 -1.055729 -2.772642 -2.37144 -1.891368
#>        [,43]     [,44]     [,45]    [,46]     [,47]     [,48]    [,49]    [,50]
#> [1,] 0.28604 -1.258561 -1.752167 11.61776 -5.828539 -2.138615 8.540101 2.523461
#>          [,51]     [,52]    [,53]    [,54]     [,55]     [,56]    [,57]
#> [1,] 0.8220901 -2.944083 3.193758 2.348336 -2.073616 -3.662581 6.140532
#>         [,58]    [,59]     [,60]    [,61]      [,62]     [,63]     [,64]
#> [1,] 4.896996 1.385081 -4.111976 3.392483 -0.1203807 -3.940615 -6.634933
#>          [,65]    [,66]     [,67]      [,68]     [,69]    [,70]    [,71]
#> [1,] 0.6176939 3.514902 -5.038842 -0.5277097 0.2546469 -1.46601 6.432335
#>          [,72]      [,73]   [,74]    [,75]    [,76]    [,77]     [,78]
#> [1,] -2.588185 -0.3736621 6.03494 2.379769 2.660543 4.683052 -1.706612
#>          [,79]    [,80]     [,81]     [,82]     [,83]      [,84]    [,85]
#> [1,] -1.270949 4.177131 -4.521285 -2.862129 -1.763723 -0.1038676 3.723349
#>          [,86]    [,87]    [,88]     [,89]     [,90]    [,91]      [,92]
#> [1,] -6.982475 1.602742 2.909348 -5.440093 -4.723676 2.095619 -0.6255857
#>         [,93]      [,94]     [,95]    [,96]    [,97]    [,98]    [,99]   [,100]
#> [1,] 4.932321 0.06629174 -8.488125 -4.03461 2.546108 2.997217 10.39723 1.724663