Integrate over Multivariate Normal Random Effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>          [,1]     [,2]    [,3]     [,4]     [,5]   [,6]     [,7]      [,8]
#> [1,] 6.094866 33.01903 -1.5904 12.62929 18.51197 12.407 15.38875 0.4148946
#>          [,9]  [,10]     [,11]    [,12]    [,13]     [,14]    [,15]    [,16]
#> [1,] 2.083895 32.543 -8.703038 -1.88836 10.75482 -9.304114 7.420675 11.06804
#>         [,17]     [,18]     [,19]     [,20]    [,21]     [,22]     [,23]
#> [1,] 9.265016 -13.59022 -3.848317 -9.944513 2.257574 -26.32581 -14.27758
#>         [,24]    [,25]     [,26]   [,27]     [,28]     [,29]     [,30]
#> [1,] 10.19999 -21.3155 -13.67737 2.50232 -14.77839 -5.864851 -15.54946
#>          [,31]     [,32]     [,33]     [,34]     [,35]     [,36]     [,37]
#> [1,] -2.477207 -16.56352 0.5974656 -25.21488 -5.580802 -20.15265 -17.81811
#>          [,38]    [,39]    [,40]     [,41]     [,42]     [,43]    [,44]
#> [1,] -14.82534 6.577652 11.92574 -1.500502 -4.655491 -11.18476 22.27759
#>          [,45]    [,46]    [,47]     [,48]    [,49]     [,50]     [,51]
#> [1,] -13.08993 24.20531 13.70225 -3.943528 16.97877 -4.246826 -9.737776
#>          [,52]     [,53]   [,54]    [,55]    [,56]    [,57]    [,58]      [,59]
#> [1,] -14.45189 -13.50489 9.62252 15.26685 6.560617 5.392053 10.90981 -0.2993726
#>          [,60]    [,61]    [,62]    [,63]    [,64]    [,65]    [,66]     [,67]
#> [1,] 0.8085449 5.610977 18.83135 8.498919 18.32375 5.772201 14.97215 -11.64173
#>          [,68]    [,69]    [,70]    [,71]    [,72]     [,73]     [,74]    [,75]
#> [1,] -3.024168 7.233206 -7.27195 -22.6957 12.20706 -12.24136 -11.83453 7.052588
#>            [,76]     [,77]    [,78]     [,79]    [,80]    [,81]     [,82]
#> [1,] -0.05077762 -1.021667 5.824326 -10.68109 12.17995 26.26156 -18.05969
#>          [,83]      [,84]    [,85]    [,86]    [,87]    [,88]  [,89]     [,90]
#> [1,] -22.16726 -0.1312907 1.496175 24.90806 23.94806 10.77637 1.8301 -17.81816
#>          [,91]    [,92]    [,93]     [,94]    [,95]     [,96]     [,97]
#> [1,] -3.590921 1.839419 33.46805 -1.437491 14.58216 -14.05199 -1.984737
#>          [,98]    [,99]    [,100]
#> [1,] -24.12695 5.582271 -7.431518

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>            [,1]     [,2]     [,3]     [,4]     [,5]      [,6]     [,7]
#> [1,] 0.08440136 -3.44775 5.088721 14.46701 8.165463 -3.820604 8.312141
#>           [,8]      [,9]     [,10]    [,11]     [,12]     [,13]      [,14]
#> [1,] -9.697645 0.7391872 -0.189629 3.870317 -3.680262 -8.616057 -0.2911991
#>         [,15]    [,16]     [,17]     [,18]     [,19]    [,20]    [,21]
#> [1,] 2.046797 2.152396 -1.575158 -2.055514 -5.133154 7.163599 3.195806
#>           [,22]     [,23]     [,24]     [,25]    [,26]    [,27]    [,28]
#> [1,] -0.5842747 -6.635836 -5.925639 -4.618883 0.142318 6.524181 12.71326
#>          [,29]    [,30]    [,31]     [,32]    [,33]    [,34]    [,35]     [,36]
#> [1,] -4.493987 1.619658 -6.51604 0.2718765 5.320845 3.582205 7.139047 0.2178829
#>         [,37]    [,38]    [,39]     [,40]    [,41]    [,42]   [,43]     [,44]
#> [1,] 4.272702 1.965904 2.764374 -1.427465 5.428538 3.920788 0.83671 -5.605008
#>         [,45]     [,46]    [,47]     [,48]     [,49]      [,50]   [,51]
#> [1,] -2.84802 -2.225961 2.961443 -5.308824 -5.874339 -0.7950069 1.44582
#>         [,52]    [,53]     [,54]     [,55]     [,56]     [,57]     [,58]
#> [1,] 0.549212 1.962279 -3.804465 -4.017516 -3.794646 0.6702188 -1.271579
#>          [,59]     [,60]    [,61]    [,62]     [,63]    [,64]    [,65]
#> [1,] -1.131699 -3.992482 5.198997 1.960652 -1.620008 4.885224 6.663421
#>          [,66]     [,67]    [,68]    [,69]     [,70]   [,71]     [,72]    [,73]
#> [1,] 0.2929297 -6.425557 2.045255 8.600357 -3.772269 4.88818 -10.54149 5.112283
#>         [,74]    [,75]      [,76]    [,77]     [,78]    [,79]     [,80]
#> [1,] 2.152271 6.589136 -0.6417778 1.993722 -8.544863 2.051894 -8.373757
#>         [,81]     [,82]    [,83]     [,84]    [,85]     [,86]    [,87]
#> [1,] 9.380383 0.7190305 3.294453 -7.427046 4.646623 0.6484273 3.551721
#>          [,88]   [,89]    [,90]     [,91]    [,92]      [,93]    [,94]    [,95]
#> [1,] 0.7570661 2.81775 1.297526 0.5641896 -6.04826 -0.9958314 1.375994 5.916953
#>            [,96]     [,97]     [,98]     [,99]    [,100]
#> [1,] -0.03328913 -6.096066 0.8635529 0.3184298 -1.330047