Integrate over Multivariate Normal Random Effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>          [,1]      [,2]      [,3]      [,4]      [,5]     [,6]     [,7]
#> [1,] -8.21975 -13.84189 -1.181137 -16.60035 -6.263398 7.824706 2.625402
#>          [,8]      [,9]     [,10]     [,11]    [,12]    [,13]    [,14]    [,15]
#> [1,] 14.27765 -12.83861 -7.865426 -12.58114 15.07451 3.219735 7.206072 7.852361
#>         [,16]    [,17]     [,18]     [,19]      [,20]    [,21]    [,22]
#> [1,] 9.377461 24.88406 -1.765076 -1.229247 -0.1259289 8.458251 -2.65909
#>         [,23]     [,24]      [,25]     [,26]    [,27]    [,28]    [,29]
#> [1,] 17.65582 -40.28105 -0.6548755 -25.46558 25.98194 9.358621 12.33937
#>         [,30]     [,31]     [,32]    [,33]     [,34]    [,35]      [,36]
#> [1,] 2.304196 -1.830956 -19.36169 -2.39621 -5.296959 6.771645 -0.9852711
#>         [,37]      [,38]     [,39]    [,40]     [,41]     [,42]     [,43]
#> [1,] 9.174793 -0.6853259 -3.214099 1.369249 -14.69451 -18.92536 -4.691912
#>          [,44]    [,45]    [,46]     [,47]     [,48]     [,49]    [,50]
#> [1,] -3.395282 5.588375 11.88789 -9.479759 -1.414884 -2.851969 23.64141
#>         [,51]    [,52]    [,53]     [,54]    [,55]     [,56]     [,57]
#> [1,] 3.956484 16.49734 17.78168 -10.60411 12.73674 -12.50116 -7.590789
#>          [,58]     [,59]     [,60]    [,61]     [,62]    [,63]     [,64]
#> [1,] -16.67301 -13.12833 -11.19637 3.015031 -19.64468 3.954813 -6.704293
#>           [,65]      [,66]    [,67]   [,68]    [,69]     [,70]     [,71]
#> [1,] -0.2509406 -0.9869485 -7.72613 -3.2416 -11.3906 -15.39388 -4.137898
#>        [,72]    [,73]    [,74]     [,75]    [,76]    [,77]   [,78]     [,79]
#> [1,] 15.6107 9.211501 -18.4155 -14.39633 12.41999 4.621457 16.3691 0.6662546
#>          [,80]      [,81]     [,82]    [,83]     [,84]    [,85]     [,86]
#> [1,] -4.420835 0.03864918 -26.33088 9.664589 -1.489762 22.41451 -5.868448
#>          [,87]    [,88]    [,89]    [,90]    [,91]    [,92]    [,93]     [,94]
#> [1,] -25.92078 6.052681 4.025393 1.048255 2.393558 13.90764 20.42656 -23.82144
#>          [,95]     [,96]     [,97]    [,98]    [,99]   [,100]
#> [1,] -4.437882 -10.02287 -31.48534 14.66136 24.79832 8.252152

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>          [,1]     [,2]      [,3]     [,4]      [,5]     [,6]     [,7]    [,8]
#> [1,] 2.147882 1.673455 -1.982376 2.594706 -3.874761 1.464272 4.015581 2.01848
#>           [,9]     [,10]     [,11]     [,12]     [,13]     [,14]      [,15]
#> [1,] -2.116144 -3.879996 -5.073983 0.2736628 -1.580126 -11.45799 0.01590322
#>         [,16]     [,17]     [,18]      [,19]      [,20]      [,21]    [,22]
#> [1,] 7.475987 -3.797318 -1.537177 -0.5051355 -0.5797352 -0.6763786 2.063976
#>         [,23]    [,24]    [,25]     [,26]     [,27]       [,28]     [,29]
#> [1,] 5.084025 5.760796 5.534571 -3.058503 -1.641095 -0.05004043 -2.669437
#>         [,30]      [,31]     [,32]    [,33]    [,34]    [,35]     [,36]
#> [1,] 1.537748 -0.4270846 -5.628215 9.722924 4.348147 8.878211 -2.635254
#>          [,37]     [,38]     [,39]     [,40]    [,41]     [,42]     [,43]
#> [1,] -3.246904 -2.870514 -4.300586 -4.357597 5.168697 -2.398616 -3.642096
#>          [,44]      [,45]    [,46]    [,47]     [,48]      [,49]    [,50]
#> [1,] -5.413545 -0.2338258 2.510791 8.060212 -6.690087 -0.7826985 7.197941
#>         [,51]     [,52]    [,53]    [,54]    [,55]     [,56]     [,57]
#> [1,] 4.446285 -7.597376 3.486963 1.461387 6.471352 -3.813873 -3.264302
#>          [,58]    [,59]     [,60]     [,61]    [,62]     [,63]    [,64]
#> [1,] 0.3833314 6.245501 -5.541226 -3.018256 3.075464 -8.349013 3.183157
#>         [,65]     [,66]    [,67]    [,68]      [,69]     [,70]    [,71]   [,72]
#> [1,] 0.513666 -1.917561 0.605156 1.882942 0.07164787 0.7987765 2.292967 8.84777
#>         [,73]     [,74]     [,75]    [,76]     [,77]    [,78]     [,79]
#> [1,] 1.829361 -7.164046 -8.823659 6.773295 -2.532076 2.816681 0.2253303
#>         [,80]     [,81]     [,82]    [,83]      [,84]     [,85]    [,86]
#> [1,] 1.860211 -4.666685 0.6436095 5.396081 -0.2365861 -1.962024 -8.17223
#>         [,87]     [,88]     [,89]     [,90]    [,91]     [,92]      [,93]
#> [1,] 10.85237 -1.466044 -2.616773 -3.298894 7.116266 -0.567621 -0.1953356
#>           [,94]     [,95]    [,96]   [,97]    [,98]     [,99]   [,100]
#> [1,] 0.06702349 -1.597441 4.898998 9.76389 1.794452 -5.983332 2.201699