Integrate over Multivariate Normal Random Effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>           [,1]     [,2]      [,3]     [,4]     [,5]      [,6]      [,7]
#> [1,] -10.73133 16.45484 -1.763922 10.85234 -2.67739 -11.61025 -15.44474
#>          [,8]     [,9]      [,10]     [,11]      [,12]     [,13]     [,14]
#> [1,] 1.348722 9.935034 -0.3748846 -9.592854 -0.2993741 -2.980853 -21.57708
#>         [,15]    [,16]    [,17]    [,18]     [,19]    [,20]   [,21]    [,22]
#> [1,] 9.454411 16.38595 9.217082 10.77465 -14.40669 32.96655 16.4079 5.381817
#>        [,23]     [,24]     [,25]    [,26]     [,27]    [,28]    [,29]    [,30]
#> [1,] 1.57902 -17.42984 -6.048337 4.687278 -4.178446 27.38421 3.423225 10.17904
#>         [,31]    [,32]    [,33]     [,34]    [,35]     [,36]    [,37]    [,38]
#> [1,] 4.353276 6.743626 9.175815 0.4050867 1.708969 -4.131945 10.83398 5.319792
#>          [,39]    [,40]    [,41]     [,42]     [,43]    [,44]    [,45]    [,46]
#> [1,] -10.46888 6.033918 11.44715 -9.628832 -22.55014 18.68625 9.288542 8.336289
#>          [,47]     [,48]     [,49]    [,50]     [,51]    [,52]     [,53]
#> [1,] -13.70786 -9.142804 0.6221032 17.37633 -5.289842 2.984856 -2.441415
#>          [,54]    [,55]    [,56]     [,57]    [,58]     [,59]      [,60]
#> [1,] -6.341261 -9.15148 16.67311 -5.480413 3.341477 -13.60777 -0.5649682
#>         [,61]    [,62]   [,63]    [,64]    [,65]    [,66]     [,67]    [,68]
#> [1,] 8.657184 25.23998 4.74047 -19.7926 3.828895 3.772589 -5.690293 3.104545
#>         [,69]    [,70]     [,71]    [,72]    [,73]    [,74]     [,75]     [,76]
#> [1,] 13.54921 -12.4339 -4.668073 16.31766 12.28719 7.613057 -30.58955 -12.60662
#>         [,77]      [,78]    [,79]     [,80]    [,81]     [,82]    [,83]
#> [1,] 13.59458 -0.2216482 9.832165 -18.19187 5.489156 0.8266161 4.573546
#>           [,84]    [,85]     [,86]   [,87]    [,88]     [,89]    [,90]
#> [1,] -0.9353907 4.544054 -4.809321 10.5014 7.076452 0.8636645 -2.22971
#>          [,91]     [,92]    [,93]    [,94]     [,95]    [,96]     [,97]
#> [1,] -4.386655 -15.00521 1.064948 8.947476 -9.045031 -3.92707 -8.784215
#>         [,98]     [,99]    [,100]
#> [1,] 2.611411 -3.231908 -17.98965

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>          [,1]      [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 5.153871 -2.889406 11.33493 7.903857 3.166639 9.867556 3.134672 5.060285
#>           [,9]    [,10]    [,11]    [,12]     [,13]    [,14]     [,15]
#> [1,] -1.292147 4.215954 6.136252 4.072892 -2.385978 2.826514 -1.926475
#>            [,16]    [,17]    [,18]    [,19]     [,20]    [,21]     [,22]
#> [1,] -0.07386512 5.701962 4.193266 7.020947 -5.154657 7.533487 -2.655129
#>          [,23]    [,24]     [,25]      [,26]   [,27]      [,28]       [,29]
#> [1,] -2.082544 7.053091 -8.492056 -0.1899815 2.86042 -0.7191079 -0.03720783
#>            [,30]     [,31]    [,32]     [,33]      [,34]     [,35]     [,36]
#> [1,] -0.08259603 -3.565721 1.339993 0.9485447 -0.2331979 -4.274189 -10.70925
#>          [,37]    [,38]    [,39]    [,40]  [,41]    [,42]    [,43]     [,44]
#> [1,] 0.0159841 9.759574 2.851816 2.365607 1.0786 3.777174 3.796825 -6.250688
#>          [,45]    [,46]     [,47]     [,48]     [,49]     [,50]    [,51]
#> [1,] -11.16837 5.868531 -2.655885 -8.359978 0.7247681 -11.45212 1.933541
#>          [,52]      [,53]    [,54]     [,55]     [,56]      [,57]      [,58]
#> [1,] -3.683082 -0.4284827 1.050095 0.3364567 -5.913375 -0.1409333 0.09047365
#>        [,59]    [,60]       [,61]     [,62]    [,63]     [,64]    [,65]
#> [1,] 9.24751 1.423398 0.005971304 -5.809248 3.228203 -2.536097 1.208887
#>         [,66]     [,67]   [,68]    [,69]    [,70]      [,71]     [,72]    [,73]
#> [1,] 1.250921 0.6124859 0.68015 3.573018 6.080599 -0.5183488 -2.303588 3.909426
#>         [,74]    [,75]    [,76]     [,77]     [,78]    [,79]     [,80]
#> [1,] 2.583352 7.145953 4.181171 -1.364336 -2.075324 6.074995 -2.432982
#>          [,81]     [,82]    [,83]    [,84]     [,85]    [,86]    [,87]    [,88]
#> [1,] -7.091163 -2.114822 4.191582 5.880913 -2.779183 5.050767 3.120203 -4.91881
#>          [,89]   [,90]      [,91]     [,92]    [,93]    [,94]    [,95]
#> [1,] -9.205767 4.91701 -0.3579858 -10.16686 6.674743 8.465742 1.349994
#>          [,96]    [,97]     [,98]   [,99]   [,100]
#> [1,] -2.229435 5.185969 -3.967376 6.03906 10.95159