Integrate over Multivariate Normal Random Effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>          [,1]      [,2]     [,3]      [,4]       [,5]     [,6]     [,7]
#> [1,] -16.1834 -3.254328 4.194579 -23.53327 -0.5188451 9.178796 4.692499
#>           [,8]      [,9]     [,10]     [,11]    [,12]    [,13]    [,14]
#> [1,] -5.920071 -5.009926 -4.277448 -5.212041 5.316278 -14.9687 -12.9002
#>         [,15]    [,16]    [,17]    [,18]    [,19]     [,20]     [,21]     [,22]
#> [1,] 8.996457 21.95123 2.400876 9.431315 32.89216 -1.879799 -23.23865 -3.647343
#>          [,23]     [,24]    [,25]    [,26]    [,27]      [,28]     [,29]
#> [1,] -12.79963 -5.925315 3.350146 18.74986 7.396753 -0.2015659 -8.046596
#>         [,30]    [,31]    [,32]    [,33]     [,34]    [,35]     [,36]     [,37]
#> [1,] 6.154662 8.764885 2.666731 5.287095 -2.407153 -1.24262 -22.62962 -20.77913
#>         [,38]     [,39]    [,40]    [,41]     [,42]    [,43]     [,44]   [,45]
#> [1,] 13.06837 -12.02255 -3.94986 20.37666 -15.82802 5.232373 -2.325956 16.3863
#>         [,46]     [,47]    [,48]    [,49]     [,50]     [,51]    [,52]
#> [1,] -19.2001 -11.07898 4.586667 -34.9371 -21.94996 -10.53565 14.19179
#>          [,53]    [,54]     [,55]     [,56]    [,57]    [,58]    [,59]    [,60]
#> [1,] -1.676121 5.969468 -5.091277 -4.135677 4.336732 2.721606 8.262858 3.053692
#>         [,61]     [,62]    [,63]    [,64]     [,65]    [,66]    [,67]     [,68]
#> [1,] 22.96605 -11.06465 27.84105 1.075863 -6.842237 2.576812 -8.75933 -15.58757
#>         [,69]     [,70]   [,71]     [,72]      [,73]     [,74]    [,75]
#> [1,] 5.993114 -4.460306 19.0297 -15.81247 -0.4233788 -25.45628 -1.45913
#>          [,76]    [,77]     [,78]    [,79]     [,80]     [,81]     [,82]
#> [1,] 0.9364227 8.068247 -6.351627 7.942667 0.8387676 -21.10884 -12.26341
#>          [,83]     [,84]     [,85]      [,86]     [,87]    [,88]     [,89]
#> [1,] -27.88381 -3.274593 -1.515602 -0.4688212 -11.03829 -5.87673 -4.379489
#>          [,90]      [,91]     [,92]     [,93]      [,94]    [,95]    [,96]
#> [1,] -9.374289 -0.3342747 -11.14268 -17.59473 -0.1603823 2.538903 -23.3764
#>         [,97]    [,98]     [,99]    [,100]
#> [1,] 5.194802 19.83621 -3.379154 -10.22109

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>          [,1]     [,2]      [,3]     [,4]     [,5]      [,6]     [,7]      [,8]
#> [1,] 1.910189 8.631617 -3.424835 3.584615 10.37838 -3.799334 6.554608 -6.443866
#>          [,9]     [,10]   [,11]     [,12]      [,13]     [,14]     [,15]
#> [1,] 1.403303 -10.37511 -3.3555 0.1190537 -0.2074201 -3.761907 -2.185345
#>         [,16]     [,17]     [,18]     [,19]     [,20]   [,21]    [,22]
#> [1,] 4.548403 -4.605914 -1.894752 -4.568833 0.3329634 -2.4507 10.29691
#>          [,23]    [,24]    [,25]     [,26]     [,27]     [,28]    [,29]
#> [1,] -2.091839 2.732583 4.472508 -9.451908 0.4504443 -4.990331 4.506836
#>         [,30]    [,31]    [,32]     [,33]     [,34]     [,35]    [,36]
#> [1,] 2.975892 3.509957 3.897699 -6.305461 -1.242186 -2.912081 4.272444
#>          [,37]     [,38]     [,39]    [,40]    [,41]    [,42]     [,43]
#> [1,] -4.647305 -1.054702 -8.028698 6.884467 1.653458 10.31134 -1.189466
#>          [,44]     [,45]    [,46]    [,47]      [,48]    [,49]     [,50]
#> [1,] -2.863625 -4.933375 4.124798 -1.47646 -0.9638359 7.129215 -5.730383
#>         [,51]     [,52]    [,53]     [,54]      [,55]    [,56]      [,57]
#> [1,] 8.280422 0.5552665 8.366015 -8.167781 -0.5267777 1.505302 -0.2428476
#>           [,58]     [,59]  [,60]     [,61]     [,62]    [,63]     [,64]
#> [1,] -0.9906218 -4.305491 -6.472 -4.438751 -8.656812 1.481124 -8.025476
#>           [,65]    [,66]     [,67]     [,68]     [,69]     [,70]    [,71]
#> [1,] -0.6043894 1.166947 -7.806392 0.6688054 0.3017386 -10.48205 2.885608
#>        [,72]    [,73]     [,74]     [,75]    [,76]    [,77]     [,78]     [,79]
#> [1,] -2.6532 -5.37225 -2.773419 -5.309871 2.788051 -5.73329 -2.198867 -1.229768
#>           [,80]    [,81]    [,82]    [,83]   [,84]     [,85]     [,86]    [,87]
#> [1,] -0.1974395 1.686664 -3.17966 5.214171 2.66407 -4.569317 -4.537142 3.627846
#>         [,88]    [,89]     [,90]      [,91]    [,92]    [,93]    [,94]
#> [1,] 3.071229 3.616029 -5.599741 -0.7085017 1.063844 5.634129 1.561238
#>          [,95]    [,96]     [,97]     [,98]     [,99]     [,100]
#> [1,] -6.095728 2.199087 -1.430702 -2.008594 -1.291112 -0.5292206