Integrate over Multivariate Normal Random Effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>           [,1]      [,2]     [,3]      [,4]      [,5]      [,6]      [,7]
#> [1,] -6.251405 -19.02388 1.903774 -17.42159 -10.61989 -14.44416 -26.90061
#>          [,8]      [,9]    [,10]     [,11]     [,12]    [,13]     [,14]   [,15]
#> [1,] 19.43799 -10.83447 3.200544 -12.86926 -5.109394 7.271884 -19.65873 22.8439
#>         [,16]    [,17]     [,18]     [,19]     [,20]    [,21]     [,22]
#> [1,] 13.19785 20.10617 -34.07144 -12.75203 -26.15217 -5.35221 -8.494622
#>         [,23]    [,24]     [,25]    [,26]     [,27]     [,28]     [,29]
#> [1,] 16.55636 10.63689 -5.861667 3.058752 -19.19603 -20.17005 0.9633702
#>          [,30]    [,31]     [,32]     [,33]     [,34]    [,35]    [,36]
#> [1,] -1.750993 1.009742 -2.315715 -3.688822 -5.239853 8.485032 10.69663
#>         [,37]     [,38]    [,39]     [,40]     [,41]     [,42]    [,43]
#> [1,] 2.355878 -4.128945 -1.14133 -19.77675 -14.14457 -2.346999 13.94545
#>          [,44]     [,45]     [,46]    [,47]     [,48]     [,49]     [,50]
#> [1,] -4.004849 -25.48371 -13.72857 11.47208 -1.021397 -11.03038 -16.46223
#>          [,51]    [,52]    [,53]     [,54]    [,55]    [,56]    [,57]    [,58]
#> [1,] -8.386473 29.51284 3.068763 -2.250651 15.70159 6.673332 7.396039 -5.37439
#>         [,59]     [,60]   [,61]      [,62]     [,63]    [,64]     [,65]
#> [1,] 3.773658 -9.553894 5.63383 -0.3412496 -2.534172 6.462263 -3.973515
#>          [,66]     [,67]     [,68]    [,69]     [,70]    [,71]     [,72]
#> [1,] 0.7607033 -5.886921 -11.02331 -24.3892 -7.286741 -7.94175 -9.720058
#>         [,73]     [,74]   [,75]    [,76]     [,77]     [,78]   [,79]    [,80]
#> [1,] 8.148411 -24.73854 10.2354 -20.2395 -7.081021 -16.96962 9.48608 2.201288
#>           [,81]    [,82]     [,83]    [,84]     [,85]    [,86]    [,87]   [,88]
#> [1,] 0.06106591 8.518086 -6.371226 10.10028 -17.87895 5.027067 12.59021 13.8288
#>            [,89]     [,90]     [,91]    [,92]     [,93]    [,94]     [,95]
#> [1,] -0.03121163 -20.60872 -4.121802 0.491061 -28.62451 3.485921 -22.37964
#>         [,96]    [,97]   [,98]     [,99]    [,100]
#> [1,] 26.84503 -2.51379 24.3283 -5.149955 -3.851023

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>          [,1]      [,2]       [,3]      [,4]      [,5]     [,6]      [,7]
#> [1,] 1.762906 -6.564966 -0.9518715 -2.953851 -2.544087 10.18274 -7.882199
#>          [,8]     [,9]    [,10]    [,11]    [,12]   [,13]    [,14]      [,15]
#> [1,] 5.670642 1.281587 0.028076 1.704286 6.304636 1.81033 9.512045 -0.3539797
#>          [,16]    [,17]     [,18]  [,19]     [,20]   [,21]     [,22]   [,23]
#> [1,] -1.955522 3.163036 0.4061787 1.7949 -2.651757 11.8849 -2.854862 5.58783
#>          [,24]  [,25]    [,26]     [,27]    [,28]    [,29]     [,30]     [,31]
#> [1,] -7.118145 2.8645 6.292171 -4.310095 11.44271 3.388143 -4.133425 -3.334267
#>         [,32]     [,33]     [,34]    [,35]    [,36]      [,37]    [,38]
#> [1,] 1.078719 -1.935436 -1.913725 4.447662 3.513345 -0.4619034 4.892971
#>         [,39]     [,40]    [,41]     [,42]    [,43]     [,44]     [,45]
#> [1,] 1.349865 -1.683508 8.326071 -3.528527 -1.30105 0.6493318 -3.789308
#>         [,46]      [,47]    [,48]    [,49]    [,50]      [,51]    [,52]
#> [1,] 1.929295 -0.1394781 1.175655 3.881204 8.583071 -0.6732932 2.309131
#>         [,53]    [,54]    [,55]      [,56]    [,57]     [,58]      [,59]
#> [1,] 1.095717 6.270732 5.738528 0.06695594 4.006492 -7.434122 -0.9000875
#>          [,60]     [,61]     [,62]    [,63]    [,64]    [,65]   [,66]     [,67]
#> [1,] -4.511593 -9.122144 -2.287459 1.159219 8.548481 6.302461 2.55317 -4.106542
#>         [,68]     [,69]     [,70]    [,71]    [,72]    [,73]     [,74]    [,75]
#> [1,] 7.448402 0.1122291 -3.208829 2.624504 1.463666 4.736407 -2.509744 1.297511
#>          [,76]    [,77]    [,78]    [,79]    [,80]    [,81]     [,82]     [,83]
#> [1,] 0.5941819 4.683491 1.121092 1.583288 7.393723 4.833735 -2.554226 -5.192848
#>         [,84]    [,85]     [,86]     [,87]    [,88]     [,89]     [,90]   [,91]
#> [1,] 4.505656 1.319812 0.0167096 -1.534787 7.891752 -4.214167 0.1152118 5.66833
#>         [,92]    [,93]    [,94]    [,95]     [,96]    [,97]   [,98]    [,99]
#> [1,] 2.841532 2.103932 2.813323 2.399921 -6.587499 2.095624 5.04571 5.695331
#>        [,100]
#> [1,] 3.773125