Integrate over Multivariate Normal Random Effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>          [,1]     [,2]     [,3]     [,4]      [,5]      [,6]     [,7]      [,8]
#> [1,] -7.13905 8.799381 5.895385 3.889209 -9.285899 -3.579989 1.041241 -8.638754
#>         [,9]    [,10]    [,11]    [,12]      [,13]   [,14]    [,15]    [,16]
#> [1,] 12.1794 12.44624 10.47257 3.941218 0.08334885 17.2786 7.036549 -30.4978
#>          [,17]    [,18]    [,19]     [,20]   [,21]     [,22]     [,23]
#> [1,] -27.44357 -2.42659 22.88125 -6.605812 -23.267 -18.29805 -17.95791
#>          [,24]     [,25]     [,26]     [,27]  [,28]    [,29]   [,30]     [,31]
#> [1,] -2.181411 -13.33407 -11.72445 -18.80181 1.2253 12.69811 4.97655 -6.704412
#>          [,32]     [,33]     [,34]     [,35]     [,36]     [,37]    [,38]
#> [1,] -10.01287 -11.45362 -11.95595 -23.29155 -18.80826 -25.94747 4.231986
#>          [,39]      [,40]    [,41]     [,42]     [,43]    [,44]    [,45]
#> [1,] -11.64863 -0.5669477 16.78906 -4.967826 -19.75819 -20.5728 4.431149
#>         [,46]    [,47]    [,48]    [,49]    [,50]     [,51]     [,52]     [,53]
#> [1,] 18.53628 2.209586 4.819415 6.528502 14.03713 -2.802122 -19.59115 -13.19098
#>         [,54]     [,55]    [,56]     [,57]    [,58]     [,59]     [,60]
#> [1,] 3.638902 -13.77974 7.271534 -2.677764 4.531194 -10.98145 -8.910231
#>         [,61]    [,62]     [,63]     [,64]     [,65]    [,66]      [,67]
#> [1,] 7.657277 25.68987 -4.923939 -12.79807 -7.718677 19.10626 -0.8990411
#>         [,68]   [,69]    [,70]    [,71]    [,72]  [,73]    [,74]     [,75]
#> [1,] -1.87426 7.44332 27.87348 1.749559 5.977769 7.0236 14.48104 -9.123079
#>         [,76]     [,77]    [,78]    [,79]     [,80]     [,81]    [,82]
#> [1,] -2.00916 -15.69062 22.21724 28.42391 0.5655805 -12.80807 7.050764
#>          [,83]      [,84]    [,85]    [,86]     [,87]    [,88]    [,89]
#> [1,] -3.267897 -0.8191023 7.388216 1.799863 -9.411718 27.96964 17.91588
#>          [,90]    [,91]    [,92]    [,93]     [,94]   [,95]     [,96]    [,97]
#> [1,] -15.39798 7.977054 3.140936 14.46796 -10.36662 11.0329 -34.25664 12.86886
#>         [,98]     [,99] [,100]
#> [1,] 10.72235 -7.172037 4.9409

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>           [,1]      [,2]     [,3]         [,4]     [,5]    [,6]      [,7]
#> [1,] -1.486617 -3.614167 5.828686 -0.001901758 -4.22139 4.68413 -5.789915
#>          [,8]       [,9]     [,10]     [,11]     [,12]     [,13]     [,14]
#> [1,] 1.222065 -0.0423126 -2.631236 -9.741264 0.5079345 0.2584584 -2.402295
#>         [,15]    [,16]   [,17]    [,18]     [,19]     [,20]    [,21]      [,22]
#> [1,] 1.504087 -2.75376 5.11495 4.496903 -14.76379 -3.013765 4.136756 -0.2067897
#>          [,23]     [,24]    [,25]      [,26]    [,27]     [,28]     [,29]
#> [1,] -4.141524 -16.20628 -1.31547 -0.5356182 4.367628 -4.798193 -3.206948
#>          [,30]    [,31]     [,32]     [,33]     [,34]    [,35]     [,36]
#> [1,] -1.073243 8.666525 -7.587282 -4.614412 -1.170054 2.037242 0.6912575
#>          [,37]    [,38]    [,39]     [,40]     [,41]    [,42]      [,43]
#> [1,] -1.534665 1.390512 5.915306 -1.264122 -2.211266 6.558854 -0.7126446
#>          [,44]    [,45]     [,46]    [,47]     [,48]     [,49]      [,50]
#> [1,] -5.476377 2.142339 -2.445604 3.107111 0.9118585 -4.093401 -0.3491898
#>         [,51]     [,52]    [,53]     [,54]     [,55]     [,56]    [,57]
#> [1,] 3.345692 -2.791556 -7.30456 -1.971627 -3.070973 -6.036242 3.352814
#>          [,58]    [,59]     [,60]     [,61]    [,62]    [,63]     [,64]
#> [1,] 0.2838475 5.334858 -5.223124 -4.444773 4.038021 9.465834 -3.594447
#>          [,65]     [,66]    [,67]      [,68]    [,69]    [,70]    [,71]
#> [1,] -1.530491 -6.821953 1.420861 -0.7743335 7.281402 1.059168 10.85732
#>           [,72]    [,73]     [,74]     [,75]     [,76]    [,77]     [,78]
#> [1,] -0.9495053 5.373671 -3.794119 -8.778771 -6.494032 5.549856 -16.28984
#>         [,79]    [,80]   [,81]    [,82]      [,83]    [,84]    [,85]    [,86]
#> [1,] 1.082805 3.644943 2.39184 1.464182 0.04683245 3.411585 4.816105 5.189681
#>          [,87]    [,88]    [,89]    [,90]     [,91]     [,92]    [,93]    [,94]
#> [1,] -7.185713 12.85351 4.450858 11.87358 -10.98819 -4.887022 1.003639 2.759328
#>         [,95]     [,96]    [,97]     [,98]     [,99]    [,100]
#> [1,] 6.936765 -3.522506 11.83874 -6.537307 0.9372611 -5.022344