Integrate over multivariate normal random effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>           [,1]     [,2]      [,3]      [,4]      [,5]     [,6]      [,7]
#> [1,] -4.988418 -19.0528 -5.802217 -13.84341 -3.581052 20.20315 -8.426494
#>           [,8]      [,9]     [,10]     [,11]    [,12]     [,13]     [,14]
#> [1,] -4.111559 -16.29401 -5.471313 -14.00574 18.79115 -10.32873 -5.757743
#>         [,15]    [,16]    [,17]    [,18]     [,19]    [,20]     [,21]     [,22]
#> [1,] 4.130392 2.616045 6.576995 6.389423 -3.140577 16.72254 -19.33116 -8.743529
#>         [,23]     [,24]    [,25]     [,26]    [,27]     [,28]    [,29]
#> [1,] 3.745063 -11.81035 -13.3067 -6.499147 8.168272 -25.55685 2.039113
#>          [,30]     [,31]     [,32]     [,33]    [,34]    [,35]     [,36]
#> [1,] -12.63015 0.8371754 -23.12768 -6.908215 5.105599 16.32293 -9.132753
#>          [,37]     [,38]    [,39]    [,40]    [,41]   [,42]    [,43]    [,44]
#> [1,] -2.324336 -11.69179 14.26454 8.779143 4.058365 16.7024 8.139191 10.45072
#>         [,45]    [,46]    [,47]     [,48]     [,49]    [,50]   [,51]     [,52]
#> [1,] 8.907804 5.485385 -3.14356 -9.388646 -12.79052 3.567389 8.78485 -3.930448
#>         [,53]     [,54]    [,55]    [,56]    [,57]     [,58]    [,59]     [,60]
#> [1,] 10.72612 -5.679846 6.271323 2.285793 -13.7262 -7.801539 7.741523 0.6154372
#>          [,61]     [,62]     [,63]      [,64]     [,65]     [,66]    [,67]
#> [1,] -29.89864 -12.02944 -25.39231 -0.3285185 -12.00505 -9.894444 3.116027
#>          [,68]    [,69]     [,70]     [,71]     [,72]     [,73]    [,74]
#> [1,] -3.940367 -1.87407 -10.89501 -29.15562 -18.09905 -18.50844 30.17209
#>         [,75]     [,76]    [,77]    [,78]    [,79]     [,80]     [,81]
#> [1,] 7.236023 -9.224872 -6.98408 32.21057 4.528493 -1.300382 -11.71267
#>          [,82]    [,83]     [,84]    [,85]      [,86]     [,87]    [,88]
#> [1,] -35.05204 10.55301 -9.238189 12.86671 -0.2778083 -16.97905 27.29599
#>         [,89]    [,90]    [,91]     [,92]     [,93]     [,94]   [,95]    [,96]
#> [1,] 39.35139 4.719696 2.939801 -12.20461 -5.147164 -1.378587 7.21267 19.62586
#>         [,97]     [,98]    [,99]   [,100]
#> [1,] 7.334381 -13.66017 5.568194 12.56037

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>          [,1]      [,2]     [,3]      [,4]      [,5]     [,6]      [,7]
#> [1,] 3.733656 -2.012989 2.721794 -3.708629 -2.872358 1.088113 -1.245786
#>          [,8]     [,9]     [,10]    [,11]    [,12]    [,13]     [,14]    [,15]
#> [1,] 11.06923 6.205777 -4.436006 14.34824 3.178691 2.335509 -3.070772 10.05356
#>          [,16]     [,17]    [,18]    [,19]      [,20]    [,21]     [,22]
#> [1,] -2.145791 -3.189118 1.141696 6.983692 -0.9636453 5.916047 0.1785772
#>         [,23]    [,24]    [,25]      [,26]    [,27]    [,28]     [,29]
#> [1,] 2.586913 6.838091 5.076704 -0.8659856 5.171817 -8.37645 -7.553098
#>          [,30]    [,31]    [,32]    [,33]    [,34]     [,35]     [,36]
#> [1,] -3.088604 1.074447 8.679361 1.994851 2.076163 -7.310618 0.5065775
#>           [,37]    [,38]   [,39]     [,40]    [,41]      [,42]     [,43]
#> [1,] -0.7257038 1.527025 3.64213 -5.618026 5.593483 -0.9706281 -3.380128
#>         [,44]     [,45]      [,46]     [,47]    [,48]    [,49]     [,50]
#> [1,] -2.43937 -3.245942 -0.9011828 0.8382901 5.791542 6.832256 -5.121482
#>          [,51]   [,52]    [,53]     [,54]      [,55]    [,56]    [,57]    [,58]
#> [1,] 0.6775853 3.18846 6.055867 -2.948737 -0.1643062 1.106066 6.245426 -1.31488
#>         [,59]     [,60]      [,61]     [,62]    [,63]      [,64]     [,65]
#> [1,] -0.44137 -6.342411 -0.4128045 -4.251941 1.684551 -0.3952292 -5.046147
#>         [,66]     [,67]    [,68]    [,69]    [,70]     [,71]      [,72]
#> [1,] 0.494968 -2.293253 4.915325 5.220594 5.675459 -5.937957 -0.3129807
#>           [,73]    [,74]    [,75]     [,76]    [,77]    [,78]    [,79]
#> [1,] -0.7793166 5.640218 12.56902 -0.373797 -5.75415 2.457824 2.512475
#>          [,80]     [,81]     [,82]     [,83]     [,84]    [,85]    [,86]
#> [1,] -2.141866 -2.861102 -9.077876 -6.235123 -1.305038 7.956116 1.620916
#>         [,87]    [,88]       [,89]    [,90]    [,91]     [,92]    [,93]
#> [1,] 2.006791 6.635265 -0.08235172 1.891732 4.331143 0.5163809 4.322593
#>         [,94]     [,95]     [,96]    [,97]    [,98]   [,99]   [,100]
#> [1,] 3.611285 -1.591504 -1.856342 7.792445 5.222364 1.21623 5.013826