Integrate over Multivariate Normal Random Effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>          [,1]     [,2]      [,3]      [,4]     [,5]    [,6]    [,7]      [,8]
#> [1,] 15.96407 3.697453 -5.787416 -1.038202 5.677732 22.6584 34.8807 -5.158601
#>           [,9]    [,10]     [,11]  [,12]     [,13]   [,14]    [,15]    [,16]
#> [1,] -1.730056 -6.82924 0.7682875 1.0809 -1.466285 3.83729 -12.2655 -4.81934
#>         [,17]     [,18]     [,19]      [,20]    [,21]   [,22]    [,23]
#> [1,] 2.622192 -13.49632 -10.45845 -0.8431214 14.18248 3.12659 16.74433
#>          [,24]     [,25]    [,26]    [,27]     [,28]    [,29]     [,30]
#> [1,] -12.26993 -10.46241 5.556433 13.20028 -23.16127 19.29022 -2.437254
#>         [,31]     [,32]    [,33]     [,34]    [,35]    [,36]    [,37]     [,38]
#> [1,] 2.349469 -4.950817 9.238059 -13.96725 11.17813 6.324754 11.16032 -9.719005
#>         [,39]     [,40]   [,41]     [,42]    [,43]     [,44]    [,45]    [,46]
#> [1,] 15.13065 -11.40887 7.89041 -10.03981 29.78614 -6.604055 5.984812 -4.88917
#>         [,47]    [,48]     [,49]    [,50]    [,51]     [,52]     [,53]
#> [1,] 18.22807 -7.93469 -17.18644 9.343171 14.53339 -32.32413 -6.641885
#>          [,54]    [,55]    [,56]    [,57]     [,58]     [,59]    [,60]    [,61]
#> [1,] -5.155906 10.34266 10.29898 17.58425 -16.93309 -37.43566 23.81251 11.41055
#>         [,62]     [,63]    [,64]    [,65]     [,66]    [,67]     [,68]    [,69]
#> [1,] -1.62564 -10.25196 3.262215 -18.2821 -16.88518 34.00932 -16.96225 12.77671
#>          [,70]   [,71]     [,72]    [,73]     [,74]    [,75]      [,76]
#> [1,] -11.04622 8.62076 -3.121284 8.678759 -5.439402 24.24851 -0.4941583
#>         [,77]    [,78]    [,79]     [,80]     [,81]   [,82]     [,83]    [,84]
#> [1,] 11.30279 47.14141 -9.61626 -7.119637 -21.59179 15.3403 -12.08248 2.320246
#>          [,85]    [,86]     [,87]     [,88]     [,89]     [,90]   [,91]
#> [1,] -27.68107 10.69548 -9.324865 -26.05894 -3.028472 -32.46837 13.8323
#>          [,92]    [,93]    [,94]    [,95]    [,96]     [,97]     [,98]
#> [1,] -4.786407 6.402194 19.78766 11.08289 -31.8653 -3.995243 -7.594224
#>          [,99]    [,100]
#> [1,] -3.749126 -7.114948

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>          [,1]      [,2]     [,3]      [,4]      [,5]      [,6]       [,7]
#> [1,] 1.281039 -2.503194 2.311048 -3.111448 -2.600153 -2.300815 -0.9977541
#>           [,8]       [,9]     [,10]      [,11]     [,12]    [,13]     [,14]
#> [1,] -7.870522 -0.6158656 -1.485471 -0.2624668 -3.291431 1.451101 -3.779618
#>         [,15]     [,16]     [,17]     [,18]     [,19]     [,20]      [,21]
#> [1,] 1.295834 -5.546061 0.3419095 -1.209847 -4.661816 -7.641606 -0.7204566
#>         [,22]    [,23]     [,24]    [,25]    [,26]   [,27]     [,28]     [,29]
#> [1,] 2.859512 10.92006 -6.590292 2.580041 3.685918 1.60497 -1.328533 -1.888863
#>           [,30]     [,31]     [,32]     [,33]    [,34]    [,35]     [,36]
#> [1,] -0.3533379 -3.166125 -1.767054 -10.37334 1.178031 -2.66025 -5.591676
#>          [,37]     [,38]     [,39]    [,40]    [,41]     [,42]    [,43]
#> [1,] -3.145819 -2.912923 -6.189365 1.845403 3.264095 -6.313027 0.157712
#>         [,44]    [,45]     [,46]     [,47]    [,48]     [,49]      [,50]
#> [1,] 6.582468 3.059635 -3.373541 -2.242055 -6.91427 -3.630519 -0.6476225
#>          [,51]     [,52]    [,53]    [,54]     [,55]     [,56]      [,57]
#> [1,] -6.850453 -4.557692 -3.09926 3.522525 -4.408483 -1.823003 -0.2165157
#>          [,58]     [,59]      [,60]    [,61]     [,62]    [,63]    [,64]
#> [1,] 0.9295465 0.7523613 -0.8134869 9.021929 -3.044772 8.135727 12.76018
#>          [,65]     [,66]    [,67]     [,68]    [,69]    [,70]     [,71]
#> [1,] -3.035283 0.6465616 2.376288 -11.93268 6.867514 3.397275 -2.064303
#>          [,72]    [,73]     [,74]    [,75]     [,76]     [,77]    [,78]
#> [1,] -1.003829 5.897365 -1.388342 3.032834 -9.046708 0.6909606 2.979953
#>         [,79]     [,80]    [,81]     [,82]     [,83]     [,84]     [,85]
#> [1,] 3.844638 -4.634065 5.453633 -3.630856 -4.224225 -1.276378 -1.682754
#>         [,86]    [,87]     [,88]      [,89]    [,90]     [,91]    [,92]
#> [1,] 5.528434 1.604393 -7.956396 -0.7054742 1.430454 -3.288515 8.612637
#>         [,93]     [,94]    [,95]    [,96]    [,97]    [,98]      [,99]   [,100]
#> [1,] 1.527094 -2.610218 8.179336 4.729734 9.906319 1.110219 -0.9457816 2.792575