Integrate over multivariate normal random effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>          [,1]     [,2]    [,3]   [,4]     [,5]     [,6]   [,7]     [,8]
#> [1,] 16.33305 -30.2223 20.9494 6.0396 -3.18892 9.115329 9.2754 3.156799
#>           [,9]     [,10]    [,11]    [,12]     [,13]    [,14]    [,15]
#> [1,] -15.06745 -12.42555 10.30447 2.968631 -7.281535 9.954517 1.119106
#>          [,16]    [,17]     [,18]     [,19]     [,20]   [,21]     [,22]
#> [1,] -2.842571 12.89652 0.2512586 -9.247797 -13.96647 20.5253 -2.927459
#>         [,23]    [,24]     [,25]     [,26]   [,27]     [,28]     [,29]
#> [1,] 6.058493 11.19446 -11.59553 -14.85627 27.5584 -3.612338 -49.30134
#>          [,30]    [,31]    [,32]     [,33]   [,34]    [,35]     [,36]    [,37]
#> [1,] -8.856914 4.764577 14.10375 -16.88056 1.49564 24.14136 -19.09635 -27.6159
#>         [,38]    [,39]     [,40]    [,41]     [,42]    [,43]    [,44]    [,45]
#> [1,] 15.98561 -11.2449 -8.250027 9.802676 -13.89188 16.90152 2.199274 21.10934
#>         [,46]    [,47]     [,48]    [,49]     [,50]     [,51]     [,52]
#> [1,] 5.873015 1.649848 -5.908131 5.225899 -2.007127 -13.98334 -5.211932
#>         [,53]    [,54]     [,55]    [,56]     [,57]    [,58]     [,59]
#> [1,] 4.215632 27.18997 -8.193767 -1.05781 -8.961285 -10.2129 -5.305237
#>          [,60]    [,61]     [,62]    [,63]     [,64]     [,65]     [,66]
#> [1,] -11.70795 3.318344 -10.74288 10.30614 -21.36618 -3.128957 -26.99235
#>         [,67]     [,68]    [,69]    [,70]    [,71]     [,72]     [,73]
#> [1,] 28.19991 -3.418237 6.451484 7.800556 7.810973 -19.35408 -4.133165
#>          [,74]     [,75]    [,76]    [,77]    [,78]     [,79]     [,80]
#> [1,] -6.205183 -3.525999 4.382581 26.82685 22.27322 -8.496148 0.2206881
#>          [,81]    [,82]    [,83]     [,84]    [,85]    [,86]   [,87]     [,88]
#> [1,] -8.784269 11.83839 3.689868 -9.690394 -22.7967 1.523843 23.3408 -22.23517
#>          [,89]    [,90]     [,91]     [,92]     [,93]     [,94]     [,95]
#> [1,] -0.558681 10.00453 -3.130178 -5.167655 -2.098994 -11.33217 -35.52031
#>          [,96]    [,97]    [,98]     [,99]   [,100]
#> [1,] -6.612151 11.01282 25.95844 -29.70709 12.74288

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>         [,1]     [,2]     [,3]     [,4]     [,5]      [,6]       [,7]      [,8]
#> [1,] 3.43281 2.636397 5.870841 2.457379 1.115525 0.3082787 -0.4702765 -6.483127
#>          [,9]     [,10]    [,11]   [,12]    [,13]     [,14]     [,15]    [,16]
#> [1,] 13.41717 0.6202849 3.591672 1.53159 13.81475 -1.065861 0.6842548 5.737355
#>         [,17]     [,18]    [,19]     [,20]     [,21]    [,22]     [,23]
#> [1,] 1.663438 -1.511094 11.18284 -1.387426 -3.498549 1.186757 -5.315515
#>          [,24]    [,25]     [,26]      [,27]   [,28]     [,29]    [,30]
#> [1,] -1.823486 3.407397 -6.606177 0.01121389 7.67208 -8.561164 0.115876
#>          [,31]    [,32]    [,33]     [,34]      [,35]     [,36]     [,37]
#> [1,] -13.96177 1.629958 4.097423 -9.246201 -0.4295137 -5.984407 -9.944979
#>         [,38]     [,39]    [,40]    [,41]    [,42]     [,43]    [,44]     [,45]
#> [1,] 8.660717 -6.881415 2.898838 1.196945 1.244148 -1.538634 5.487904 -1.315356
#>         [,46]     [,47]    [,48]    [,49]     [,50]    [,51]    [,52]    [,53]
#> [1,] 6.539885 -5.757777 3.416981 6.195289 -4.584654 2.244665 -1.39521 5.841204
#>          [,54]     [,55]    [,56]     [,57]    [,58]    [,59]     [,60]
#> [1,] -3.504928 -2.102293 1.433335 -1.966759 8.983297 7.703187 -11.77789
#>         [,61]     [,62]    [,63]    [,64]    [,65]      [,66]     [,67]   [,68]
#> [1,] 8.931888 -2.386485 2.203727 3.596821 1.850673 -0.7869638 -2.777494 2.06051
#>         [,69]    [,70]    [,71]    [,72]    [,73]    [,74]      [,75]    [,76]
#> [1,] 7.657555 6.308244 1.018075 2.920892 1.558166 1.067686 -0.6207758 6.257532
#>         [,77]    [,78]     [,79]     [,80]     [,81]     [,82]     [,83]
#> [1,] 3.357383 2.556622 -1.527602 -5.948294 -11.38862 -5.098328 -1.275553
#>           [,84]      [,85]     [,86]    [,87]     [,88]     [,89]    [,90]
#> [1,] 0.02470892 0.03810666 -2.372223 2.676707 -2.514837 -4.021355 4.730929
#>         [,91]    [,92]     [,93]     [,94]     [,95]    [,96]    [,97]
#> [1,] 10.90045 -0.63943 0.5314539 -2.774369 -3.167498 -5.16074 3.433273
#>          [,98]     [,99]   [,100]
#> [1,] -2.940599 -3.523969 14.62556