Integrate over Multivariate Normal Random Effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>           [,1]      [,2]     [,3]     [,4]      [,5]     [,6]     [,7]     [,8]
#> [1,] -1.966272 -7.079572 7.117384 -1.32985 -12.98731 8.131683 9.485049 2.376841
#>           [,9]     [,10]     [,11]     [,12]     [,13]     [,14]     [,15]
#> [1,] -5.255732 -8.645475 -9.053795 -1.325012 -16.50034 0.4361463 -7.906185
#>          [,16]    [,17]      [,18]     [,19]     [,20]    [,21]    [,22]
#> [1,] -26.84691 -5.92409 -0.1388985 -11.66136 -15.13183 2.792236 14.72142
#>         [,23]    [,24]      [,25]     [,26]      [,27]    [,28]    [,29]
#> [1,] 4.952927 8.393917 -0.9800322 0.8750243 -0.3465941 18.07209 2.156916
#>          [,30]    [,31]    [,32]     [,33]   [,34]    [,35]   [,36]    [,37]
#> [1,] -1.226293 13.19439 7.386154 -5.266504 13.1928 7.676028 -11.189 9.706474
#>         [,38]    [,39]    [,40]     [,41]     [,42]    [,43]    [,44]    [,45]
#> [1,] 18.94127 10.20993 4.148915 -3.426121 -6.881561 1.090512 2.953652 6.958723
#>          [,46]     [,47]    [,48]     [,49]     [,50]     [,51]     [,52]
#> [1,] -4.704731 -3.572965 2.266609 -2.684659 -7.917922 0.9749016 -13.40299
#>         [,53]     [,54]     [,55]   [,56]    [,57]    [,58]     [,59]    [,60]
#> [1,] 10.81213 -2.272833 -5.512363 18.3823 -5.48627 13.76357 -3.434739 9.472102
#>        [,61]    [,62]     [,63]     [,64]     [,65]     [,66]    [,67]    [,68]
#> [1,] 25.5854 12.13075 -12.24252 -15.23165 -4.709152 -2.387573 9.982093 2.115824
#>          [,69]    [,70]  [,71]     [,72]    [,73]    [,74]   [,75]    [,76]
#> [1,] -2.604826 18.83632 2.1809 -7.014224 8.148176 18.06477 12.8492 12.47181
#>         [,77]    [,78]     [,79]     [,80]     [,81]    [,82]    [,83]
#> [1,] 11.34787 17.50983 -1.158137 -15.24255 -14.20393 8.452221 7.517511
#>          [,84]    [,85]    [,86]    [,87]    [,88]     [,89]     [,90]
#> [1,] -11.47842 15.45604 13.60625 3.401512 4.557163 -4.539154 -6.962666
#>          [,91]     [,92]     [,93]     [,94]     [,95]     [,96]     [,97]
#> [1,] -18.52006 -18.71833 -12.75531 -17.41263 -6.051537 -23.18961 -16.45175
#>         [,98]     [,99]    [,100]
#> [1,] 7.575247 -1.362192 -16.82008

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>          [,1]      [,2]     [,3]      [,4]      [,5]       [,6]     [,7]
#> [1,] 1.741123 -2.517838 2.052935 -1.207595 -5.355822 -0.9013812 2.573862
#>          [,8]      [,9]     [,10]     [,11]     [,12]    [,13]    [,14]
#> [1,] 5.671636 -4.003895 -6.693951 -9.860532 -1.321866 6.378243 2.264152
#>         [,15]     [,16]     [,17]    [,18]      [,19]     [,20]     [,21]
#> [1,] 3.743285 -1.454634 0.7002156 3.181158 -0.5004942 -5.402999 0.5692685
#>         [,22]   [,23]   [,24]     [,25]     [,26]     [,27]     [,28]     [,29]
#> [1,] 6.320747 -2.5299 3.43556 -8.282226 -9.424544 -2.480817 -1.075035 -8.742446
#>        [,30]     [,31]     [,32]     [,33]    [,34]    [,35]     [,36]    [,37]
#> [1,] 3.44547 -9.330181 -4.336144 -1.371508 2.323085 9.642061 -2.053261 4.823427
#>          [,38]      [,39]     [,40]    [,41]     [,42]      [,43]     [,44]
#> [1,] -2.297964 -0.9056532 0.9929671 7.382101 -3.969109 -0.2811916 -2.800946
#>         [,45]     [,46]   [,47]      [,48]    [,49]    [,50]     [,51]    [,52]
#> [1,] 4.521585 -3.095447 -4.7097 -0.3635601 4.150703 2.164332 -4.108612 -12.7497
#>         [,53]    [,54]     [,55]     [,56]    [,57]     [,58]    [,59]    [,60]
#> [1,] 3.798646 3.121028 -12.42302 -3.679099 3.300518 -1.594868 0.695643 -8.68698
#>          [,61]     [,62]     [,63]     [,64]     [,65]     [,66]    [,67]
#> [1,] -2.689905 -2.000787 -15.49387 -5.235894 -7.927915 -4.441366 5.360633
#>          [,68]    [,69]    [,70]    [,71]     [,72]       [,73]     [,74]
#> [1,] -5.261679 4.605485 3.418128 -2.95496 -10.12853 -0.03465268 -6.726422
#>          [,75]     [,76]     [,77]    [,78]    [,79]      [,80]    [,81]
#> [1,] -7.731846 -5.091639 0.9594012 10.09937 6.790616 -0.6471928 3.038482
#>         [,82]     [,83]     [,84]     [,85]      [,86]    [,87]    [,88]
#> [1,] 2.118873 -7.202249 -8.391716 -1.562209 -0.3986824 2.837655 1.258095
#>         [,89]     [,90]   [,91]    [,92]    [,93]     [,94]     [,95]    [,96]
#> [1,] 6.246401 -8.052803 4.31139 6.360824 1.272269 -12.12434 -3.763075 3.665098
#>         [,97]     [,98]    [,99]    [,100]
#> [1,] 3.941549 -5.393698 3.846298 0.3557625