Integrate over Multivariate Normal Random Effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>         [,1]      [,2]     [,3]    [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 15.9004 -20.44472 12.73116 3.83661 5.967106 21.48775 20.05435 5.199038
#>         [,9]    [,10]     [,11]    [,12]     [,13]    [,14]    [,15]     [,16]
#> [1,] 9.69701 8.238271 -13.27457 5.931651 -2.745272 6.646238 1.369388 -1.604061
#>           [,17]     [,18]     [,19]      [,20]    [,21]     [,22]     [,23]
#> [1,] -0.5910549 -2.922583 -2.227293 -0.9557137 8.968595 -20.79818 -2.634153
#>         [,24]     [,25]      [,26]     [,27]     [,28]    [,29]     [,30]
#> [1,] 7.735421 -16.30389 -0.6482397 -17.62907 -10.97339 -4.32472 -9.978245
#>          [,31]    [,32]    [,33]     [,34]    [,35]    [,36]     [,37]    [,38]
#> [1,] -36.37827 16.26792 7.463791 -1.149394 11.10381 4.694845 -4.968448 1.079915
#>          [,39]     [,40]    [,41]    [,42]    [,43]     [,44]    [,45]
#> [1,] -18.03411 -9.715656 16.11453 -17.0116 19.75844 -7.878535 -7.24102
#>          [,46]    [,47]    [,48]     [,49]    [,50]    [,51]     [,52]
#> [1,] -37.97283 15.70675 13.91758 -2.494052 5.105721 31.26184 -8.418164
#>          [,53]       [,54]     [,55]    [,56]    [,57]    [,58]     [,59]
#> [1,] -10.67114 -0.08855286 -4.573378 4.238267 6.674546 6.024019 -6.121048
#>         [,60]     [,61]    [,62]    [,63]    [,64]     [,65]    [,66]    [,67]
#> [1,] 3.665311 0.9948729 8.107427 9.239574 4.227556 -21.02316 19.43458 13.51149
#>          [,68]     [,69]    [,70]    [,71]     [,72]     [,73]    [,74]
#> [1,] -1.745939 -11.12271 5.470133 3.188649 -2.193241 0.3812205 14.18896
#>         [,75]   [,76]      [,77]    [,78]     [,79]     [,80]     [,81]
#> [1,] -30.8278 11.7095 -0.7250913 3.351235 -1.966091 -5.916923 -3.808868
#>          [,82]    [,83]    [,84]     [,85]     [,86]    [,87]    [,88]    [,89]
#> [1,] -15.86057 5.610756 23.18626 -4.412192 -8.557978 2.110747 17.92548 7.607103
#>         [,90]     [,91]     [,92]    [,93]    [,94]    [,95]    [,96]     [,97]
#> [1,] 4.589687 -39.31888 -27.13433 32.53508 25.70513 18.06393 8.947948 -4.549587
#>         [,98]     [,99]   [,100]
#> [1,] 33.12638 -14.63124 10.43329

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>           [,1]     [,2]      [,3]      [,4]     [,5]       [,6]       [,7]
#> [1,] 0.9236454 8.095694 -2.361251 -1.063174 6.714794 -0.6265589 0.01047386
#>           [,8]     [,9]   [,10]   [,11]    [,12]       [,13]     [,14]    [,15]
#> [1,] -8.823251 1.358407 8.67582 1.18236 4.500019 -0.04509557 -2.705873 5.598428
#>          [,16]     [,17]    [,18]        [,19]    [,20]     [,21]   [,22]
#> [1,] -3.802668 -9.149401 6.383591 -0.006075884 5.018387 -2.102223 -1.5227
#>          [,23]      [,24]   [,25]     [,26]     [,27]     [,28]     [,29]
#> [1,] -4.976674 -0.4981036 4.93686 0.7287533 0.2715107 -1.385157 -5.098854
#>          [,30]     [,31]      [,32]     [,33]     [,34]    [,35]     [,36]
#> [1,] -3.668647 0.3438559 -0.1812282 0.9784713 -3.255337 7.935307 -4.315666
#>         [,37]     [,38]      [,39]     [,40]    [,41]    [,42]    [,43]
#> [1,] 3.736784 -2.827868 -0.1056804 -3.021924 8.717653 1.349974 1.495766
#>         [,44]    [,45]     [,46]    [,47]     [,48]   [,49]     [,50]    [,51]
#> [1,] -3.32382 8.420384 -4.403603 2.587449 -4.828995 2.45071 -4.257126 -4.17505
#>          [,52]     [,53]     [,54]    [,55]     [,56]    [,57]    [,58]
#> [1,] -6.728297 -6.026252 -3.311802 3.588431 -5.229676 6.543125 1.684492
#>         [,59]    [,60]     [,61]    [,62]    [,63]    [,64]    [,65]    [,66]
#> [1,] 1.674386 0.223584 -6.164659 2.046775 4.744272 8.203476 3.128273 4.750732
#>          [,67]     [,68]    [,69]    [,70]    [,71]     [,72]      [,73]
#> [1,] 0.9853826 0.6747663 3.252568 -1.22738 1.158707 0.4691716 -0.7283285
#>          [,74]     [,75]     [,76]     [,77]    [,78]    [,79]     [,80]
#> [1,] -1.293147 -4.003704 -3.245994 -1.298654 3.931704 16.42207 -5.059815
#>         [,81]     [,82]     [,83]    [,84]     [,85]    [,86]     [,87]
#> [1,] -1.43905 -1.613066 -1.790554 -7.53369 0.8321399 -6.55666 0.8658179
#>           [,88]      [,89]     [,90]    [,91]     [,92]     [,93]     [,94]
#> [1,] -0.3497112 -0.2083277 -11.24974 1.784301 -3.777167 0.3698163 -8.931618
#>         [,95]    [,96]    [,97]     [,98]    [,99]    [,100]
#> [1,] 4.269529 7.130581 4.225934 -3.519742 5.334583 -9.181747