Integrate over multivariate normal random effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>          [,1]      [,2]      [,3]     [,4]      [,5]     [,6]      [,7]
#> [1,] 5.508919 -9.046428 -10.81853 -9.72506 -4.600669 27.36811 -22.39579
#>           [,8]      [,9]     [,10]    [,11]    [,12]     [,13]    [,14]
#> [1,] -13.30185 -24.54852 -13.00437 12.94352 8.060632 -13.00505 19.93708
#>         [,15]     [,16]    [,17]     [,18]     [,19]    [,20]    [,21]
#> [1,] -17.5869 -11.28831 22.52933 -9.457644 -6.201337 7.423805 7.144486
#>          [,22]     [,23]     [,24]    [,25]     [,26]      [,27]    [,28]
#> [1,] -13.12711 -7.131277 -6.769451 15.77856 -4.333127 -0.5866975 -1.09863
#>          [,29]    [,30]     [,31]    [,32]     [,33]      [,34]     [,35]
#> [1,] -11.02576 1.266714 -9.476059 34.68818 -12.61915 -0.3195171 -9.005034
#>         [,36]     [,37]    [,38]    [,39]    [,40]    [,41]    [,42]    [,43]
#> [1,] 20.43119 -5.241799 21.38899 15.64738 1.630152 1.158342 15.89334 23.73674
#>         [,44]     [,45]    [,46]    [,47]     [,48]    [,49]     [,50]    [,51]
#> [1,] 1.037624 -5.785231 23.59053 2.180857 -3.460776 -15.7569 -18.09674 19.72337
#>         [,52]    [,53]     [,54]     [,55]     [,56]     [,57]    [,58]
#> [1,] 10.29978 5.387526 0.5829736 -9.918186 -2.347559 -24.97599 23.42174
#>         [,59]   [,60]   [,61]     [,62]   [,63]    [,64]    [,65]     [,66]
#> [1,] 22.62167 8.15256 9.35229 -1.472195 3.67809 3.616595 3.857996 -1.980044
#>         [,67]     [,68]    [,69]    [,70]    [,71]     [,72]   [,73]     [,74]
#> [1,] 18.48878 -8.865523 4.955395 0.142429 28.55869 -20.50089 1.60731 -4.024326
#>          [,75]    [,76]     [,77]    [,78]     [,79]    [,80]     [,81]
#> [1,] 0.7629076 16.73645 -2.523993 15.61611 -5.783767 15.47491 -11.35399
#>         [,82]     [,83]     [,84]    [,85]    [,86]    [,87]    [,88]     [,89]
#> [1,] 22.63275 -37.57641 0.3651981 20.06995 15.39677 2.283635 3.709966 -4.612179
#>          [,90]    [,91]    [,92]     [,93]     [,94]    [,95]    [,96]
#> [1,] -26.00357 21.67589 25.39133 -5.254331 -9.732676 12.94066 4.219745
#>          [,97]    [,98]     [,99]    [,100]
#> [1,] -7.847753 8.015525 -9.712482 -10.71015

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>          [,1]      [,2]     [,3]      [,4]     [,5]     [,6]    [,7]     [,8]
#> [1,] -2.13786 -1.141886 1.471798 -7.123332 10.57957 12.27216 6.62866 5.081285
#>           [,9]     [,10]    [,11]     [,12]      [,13]     [,14]     [,15]
#> [1,] -0.751992 -6.336366 1.644795 -8.975804 -0.8619069 -1.437912 -3.222835
#>          [,16]     [,17]    [,18]   [,19]     [,20]    [,21]    [,22]     [,23]
#> [1,] -4.855124 0.2871588 1.714992 6.94774 -4.272629 -7.70616 -3.04165 -4.159636
#>          [,24]   [,25]     [,26]    [,27]     [,28]     [,29]     [,30]
#> [1,] -10.86316 5.53071 -8.986518 1.059587 -5.021276 -1.093694 0.2099363
#>         [,31]     [,32]     [,33]     [,34]    [,35]    [,36]    [,37]
#> [1,] 2.010072 0.1252251 -8.357394 -6.042602 5.156604 9.501253 2.134353
#>          [,38]     [,39]      [,40]    [,41]    [,42]    [,43]    [,44]
#> [1,] -3.139673 -1.315909 -0.4879144 3.216724 -1.54102 1.925825 6.504274
#>          [,45]     [,46]    [,47]    [,48]    [,49]     [,50]   [,51]     [,52]
#> [1,] -6.081993 -9.425004 1.448699 -5.22656 3.828066 -7.045905 1.54112 -5.581919
#>          [,53]    [,54]   [,55]     [,56]      [,57]    [,58]    [,59]
#> [1,] 0.6738713 9.976221 1.90321 -1.399954 -0.7431194 2.144489 4.816868
#>          [,60]     [,61]    [,62]    [,63]     [,64]     [,65]     [,66]
#> [1,] -6.629416 -6.206007 3.556803 1.273525 -9.064631 -3.469977 -3.685034
#>          [,67]    [,68]     [,69]    [,70]     [,71]    [,72]     [,73]
#> [1,] -2.959108 1.573309 -7.685685 1.573472 -9.456442 6.244862 -2.105016
#>         [,74]    [,75]     [,76]     [,77]     [,78]     [,79]   [,80]    [,81]
#> [1,] 8.819748 6.273248 -1.293587 -10.70922 0.2830552 -0.352409 3.01338 4.843446
#>          [,82]    [,83]     [,84]    [,85]    [,86]    [,87]      [,88]
#> [1,] -4.342015 5.399338 -1.650662 -8.60626 7.402258 9.773545 -0.1484917
#>          [,89]    [,90]    [,91]    [,92]    [,93]    [,94]     [,95]    [,96]
#> [1,] -7.472259 9.035272 2.605539 -7.81383 -6.56198 4.101079 -3.493636 4.887055
#>         [,97]     [,98]    [,99]    [,100]
#> [1,] 3.895852 0.3751104 9.153939 -3.853592