Integrate over multivariate normal random effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>          [,1]      [,2]      [,3]     [,4]      [,5]      [,6]     [,7]
#> [1,] 17.97935 -14.38639 -13.03568 32.28638 -3.154047 -18.82786 15.94701
#>          [,8]      [,9]      [,10]    [,11]    [,12]    [,13]    [,14]
#> [1,] 2.736952 0.2110601 -0.7030288 16.70081 2.202977 -26.5233 -24.2142
#>          [,15]    [,16]    [,17]    [,18]    [,19]     [,20]    [,21]     [,22]
#> [1,] -10.04558 8.729764 -9.12041 9.168713 7.267786 -18.99373 11.85236 -10.80177
#>          [,23]     [,24]      [,25]     [,26]     [,27]    [,28]    [,29]
#> [1,] -11.89766 -5.410605 -0.1320495 -8.898047 -6.695073 4.077465 -2.43325
#>         [,30]    [,31]    [,32]    [,33]    [,34]      [,35]    [,36]     [,37]
#> [1,] -14.0186 1.360461 6.799071 19.00458 6.867057 -0.7566864 12.49413 -7.254549
#>         [,38]     [,39]     [,40]     [,41]    [,42]     [,43]     [,44]
#> [1,] 2.377602 -3.998716 -15.90342 0.8254158 5.534025 -12.77282 -1.030569
#>          [,45]     [,46]     [,47]     [,48]     [,49]   [,50]    [,51]
#> [1,] 0.9105706 -15.85652 -11.37826 -2.712244 -2.474723 1.19715 7.880237
#>         [,52]    [,53]     [,54]     [,55]     [,56]    [,57]     [,58]
#> [1,] 6.136016 9.783984 -2.536788 -25.19224 -4.156134 1.347544 -3.141469
#>         [,59]    [,60]     [,61]    [,62]     [,63]     [,64]   [,65]   [,66]
#> [1,] 2.017918 7.299258 -10.85498 13.29029 -7.226878 -28.29438 13.8299 -1.4357
#>         [,67]     [,68]     [,69]    [,70]    [,71]     [,72]    [,73]    [,74]
#> [1,] 12.09335 -11.91227 -5.072785 3.238745 11.62435 -15.81306 8.200748 22.61505
#>         [,75]     [,76]    [,77]     [,78]    [,79]     [,80]     [,81]
#> [1,] 4.076423 -16.37847 5.612331 -26.67344 8.505239 -14.81988 -5.208501
#>         [,82]     [,83]    [,84]    [,85]     [,86]   [,87]    [,88]    [,89]
#> [1,] -9.71307 -5.299321 4.098058 7.813956 -1.809701 18.3341 29.19979 10.09515
#>         [,90]    [,91]      [,92]     [,93]    [,94]    [,95]    [,96]    [,97]
#> [1,] 5.983667 10.77094 -0.3226981 -10.26402 3.338543 5.559697 4.760983 13.11837
#>         [,98]     [,99]   [,100]
#> [1,] 7.061265 -17.01664 4.533474

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>           [,1]      [,2]      [,3]     [,4]     [,5]      [,6]     [,7]
#> [1,] -3.458484 -3.224937 -5.141747 -3.54114 10.42086 0.8941973 4.584831
#>           [,8]     [,9]     [,10]     [,11]    [,12]      [,13]      [,14]
#> [1,] -11.60661 -7.56344 0.3741844 0.8023412 3.270631 -0.2994746 -0.7226602
#>         [,15]    [,16]     [,17]      [,18]     [,19]   [,20]     [,21]
#> [1,] -2.68012 5.101286 -5.284716 -0.5483631 -2.156125 1.17202 -12.34992
#>          [,22]       [,23]    [,24]     [,25]    [,26]      [,27]      [,28]
#> [1,] -3.020107 -0.09445924 3.190949 -1.868959 1.241598 -0.5498577 -0.8094702
#>         [,29]     [,30]    [,31]     [,32]     [,33]      [,34]    [,35]
#> [1,] 3.893272 -3.509048 2.998823 -1.306315 0.3162425 -0.4188199 1.945058
#>         [,36]     [,37]    [,38]     [,39]     [,40]    [,41]     [,42]
#> [1,] -6.35667 -3.926765 3.657213 -4.096764 -4.054458 2.691711 -2.246756
#>          [,43]    [,44]     [,45]    [,46]   [,47]    [,48]     [,49]    [,50]
#> [1,] -7.369612 -5.68378 -2.874209 10.12583 8.65002 3.384559 -8.179006 -2.61378
#>          [,51]     [,52]   [,53]      [,54]    [,55]    [,56]     [,57]
#> [1,] -7.795684 -3.591627 6.16772 -0.5007183 2.464638 8.493204 -2.101572
#>         [,58]     [,59]     [,60]     [,61]      [,62]     [,63]    [,64]
#> [1,] 7.878466 -10.18731 0.7818922 -5.507335 -0.4081805 -1.761431 5.229219
#>           [,65]    [,66]     [,67]   [,68]    [,69]     [,70]    [,71]    [,72]
#> [1,] -0.3233288 3.603097 -8.693776 3.05362 4.777086 -6.203591 2.203525 2.571237
#>        [,73]    [,74]    [,75]     [,76]    [,77]    [,78]    [,79]     [,80]
#> [1,] -1.6801 7.184514 4.996022 -1.140429 1.484745 7.313009 4.233772 -1.678533
#>         [,81]       [,82]     [,83]     [,84]     [,85]     [,86]     [,87]
#> [1,] 8.569691 -0.06767587 -6.662609 -2.994345 -1.464332 -7.630842 -3.553023
#>         [,88]     [,89]     [,90]    [,91]     [,92]     [,93]     [,94]
#> [1,] -2.23884 -7.809365 -6.046763 2.540982 -4.238702 -2.027538 -5.492847
#>         [,95]   [,96]     [,97]      [,98]   [,99]      [,100]
#> [1,] 1.160921 6.70714 -1.939315 -0.5759736 7.97216 -0.01892449