Integrate over multivariate normal random effects

Used in the process of Monte Carlo integration over multivariate normal random effects. This generates the random draws from the multivariate normal distribution and multiplies these by the data. Not intended to be called directly by most users.

Usage

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X

A numeric matrix of the data to be multiplied by the random effects

k

An integer, the number of random samples to be used for numerical integration

sd

A numeric vector of the standard deviations

chol

A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(
  X = matrix(1, 1, 2),
  k = 100L,
  sd = c(10, 5),
  chol = chol(matrix(c(1, .5, .5, 1), 2)))
#>          [,1]      [,2]     [,3]     [,4]      [,5]      [,6]     [,7]
#> [1,] 19.09336 -7.479968 19.58102 8.219261 -33.54135 -16.34742 8.195922
#>           [,8]   [,9]    [,10]    [,11]    [,12]    [,13]    [,14]     [,15]
#> [1,] -1.768265 26.815 9.700449 4.910351 28.02609 5.536787 18.28424 -1.154537
#>          [,16]     [,17]     [,18]   [,19]    [,20]     [,21]    [,22]    [,23]
#> [1,] -14.27779 -12.27549 -1.898693 11.5198 8.407076 -11.15316 -31.5484 13.85937
#>          [,24]     [,25]     [,26]     [,27]    [,28]    [,29]     [,30]
#> [1,] -28.76852 -2.849661 -18.03569 -2.689359 8.218042 18.93852 -11.56226
#>         [,31]     [,32]     [,33]     [,34]     [,35]     [,36]    [,37]
#> [1,] 23.95813 -1.974177 -22.44025 -2.618107 -18.97941 -6.784481 9.029377
#>         [,38]    [,39]     [,40]     [,41]     [,42]     [,43]    [,44]
#> [1,] 15.80025 4.195687 -31.11016 -8.943839 -12.39195 -4.638266 7.380817
#>         [,45]    [,46]    [,47]    [,48]     [,49]     [,50]    [,51]     [,52]
#> [1,] 13.23929 4.940184 12.42338 5.086007 -7.390549 -3.704462 16.86753 -8.357364
#>        [,53]     [,54]     [,55]    [,56]    [,57]     [,58]     [,59]    [,60]
#> [1,] 21.1351 -16.38413 -1.005092 2.402486 9.428074 -13.08771 -5.065737 5.974813
#>          [,61]   [,62]     [,63]     [,64]     [,65]    [,66]     [,67]
#> [1,] -6.136144 6.88458 -6.659198 -16.46153 -5.774339 3.866492 -10.54101
#>          [,68]   [,69]     [,70]     [,71]     [,72]    [,73]     [,74]
#> [1,] -5.394169 6.05784 -25.89562 -27.91229 -13.96445 -10.5158 -9.465717
#>          [,75]     [,76]    [,77]     [,78]     [,79]   [,80]    [,81]    [,82]
#> [1,] -12.98452 -19.81099 3.923423 -2.432679 -12.17246 29.5492 6.914434 -12.6909
#>         [,83]    [,84]   [,85]    [,86]     [,87]    [,88]   [,89]    [,90]
#> [1,] 1.106355 -9.35921 13.5603 6.162423 0.5531064 15.75049 14.0693 19.13502
#>         [,91]     [,92]    [,93]   [,94]    [,95]     [,96]    [,97]
#> [1,] 10.37766 -4.753662 11.94609 6.75097 3.082684 -9.425774 7.655539
#>            [,98]     [,99]    [,100]
#> [1,] -0.01170762 -1.593793 -37.99513

integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>          [,1]     [,2]       [,3]     [,4]     [,5]     [,6]    [,7]      [,8]
#> [1,] 2.963011 -12.0813 -0.2663274 3.834038 5.246043 -5.59121 9.59055 0.2816843
#>        [,9]    [,10]     [,11]     [,12]    [,13]     [,14]    [,15]    [,16]
#> [1,] 3.0551 4.846897 -2.425375 -3.951665 3.095529 -9.040826 3.250997 8.281087
#>          [,17]     [,18]    [,19]     [,20]     [,21]     [,22]     [,23]
#> [1,] 0.4092555 -3.356238 15.53746 -4.113635 -9.511849 -6.318694 -8.915466
#>         [,24]     [,25]     [,26]      [,27]     [,28]    [,29]     [,30]
#> [1,] 1.116371 -2.678878 -2.846808 -0.4417824 -2.291416 6.841759 -16.52675
#>         [,31]     [,32]     [,33]    [,34]     [,35]     [,36]    [,37]
#> [1,] 8.087682 -5.629102 -2.170906 1.646978 -2.755873 -3.059817 3.048326
#>           [,38]     [,39]     [,40]     [,41]     [,42]    [,43]    [,44]
#> [1,] -0.5470741 0.8734567 -3.454488 -5.038196 0.4860459 -5.80294 4.212368
#>         [,45]     [,46]     [,47]     [,48]     [,49]     [,50]     [,51]
#> [1,] 7.984381 0.5929219 -0.646718 -5.574264 0.5470369 0.3806081 -6.151585
#>         [,52]    [,53]    [,54]     [,55]     [,56]     [,57]     [,58]
#> [1,] 4.829295 9.898235 6.027621 0.7685779 -5.707558 -3.804394 -8.083089
#>          [,59]     [,60]     [,61]     [,62]    [,63]   [,64]    [,65]
#> [1,] -1.449136 -3.848589 -2.070678 0.4215571 3.409358 7.11971 2.680189
#>          [,66]    [,67]     [,68]     [,69]     [,70]    [,71]    [,72]
#> [1,] -3.206973 8.776148 -5.016579 -7.116154 -2.445507 3.190087 12.96551
#>          [,73]     [,74]    [,75]   [,76]    [,77]     [,78]    [,79]    [,80]
#> [1,] -5.143438 -3.253367 3.296754 1.53179 7.466713 -4.666231 5.731755 -2.76286
#>          [,81]     [,82]   [,83]    [,84]     [,85]    [,86]    [,87]    [,88]
#> [1,] 0.5479889 -1.648472 9.44422 3.261664 -6.946071 3.628199 1.311474 2.791595
#>         [,89]    [,90]     [,91]     [,92]      [,93]     [,94]   [,95]
#> [1,] 2.534597 2.565022 -4.341993 -7.814438 -0.7878935 -7.790294 9.88349
#>         [,96]     [,97]     [,98]     [,99]    [,100]
#> [1,] 5.075037 0.3529236 -1.174389 -2.859615 -5.456626