Integrate over multivariate normal random effects

TODO: write description.

integratemvn(X, k, sd, chol)

integratemvnR(X, k, sd, chol)

Arguments

X
A numeric matrix of the data to be multiplied by the random effects

k
An integer, the number of random samples to be used for numerical integration

sd
A numeric vector of the standard deviations

chol
A numeric matrix, which should be the Cholesky decomposition of the correlation matrix of the multivariate normal distribution.

Value

A numeric matrix with random values

Functions

integratemvnR: Pure R implementation of integratemvn

Examples

integratemvn(matrix(1, 1, 2), 100L, c(10, 5), chol(matrix(c(1, .5, .5, 1), 2)))
#>          [,1]     [,2]     [,3]      [,4]      [,5]     [,6]       [,7]
#> [1,] 5.820763 3.358061 17.69549 -2.929731 -3.016835 22.47545 -0.8303446
#>          [,8]     [,9]    [,10]     [,11]     [,12]    [,13]     [,14]    [,15]
#> [1,] 12.43875 10.86663 19.36848 -11.02877 -3.863258 16.18339 -13.80339 20.24105
#>          [,16]     [,17]     [,18]     [,19]    [,20]     [,21]     [,22]
#> [1,] -22.47897 -13.93332 -17.93254 -4.630021 6.628668 -0.808501 -5.855097
#>         [,23]    [,24]     [,25]     [,26]     [,27]    [,28]    [,29]
#> [1,] 4.411399 4.910083 -6.955584 -21.32713 -23.31158 21.25017 11.16589
#>          [,30]    [,31]     [,32]     [,33]     [,34]   [,35]   [,36]   [,37]
#> [1,] -9.053557 -39.3175 -20.18126 -29.51268 -6.153077 7.23051 1.27528 6.24862
#>          [,38]     [,39]     [,40]     [,41]     [,42]     [,43]     [,44]
#> [1,] 0.5254274 -18.78564 -18.59253 -8.628483 -2.007753 -1.248557 -10.73638
#>          [,45]     [,46]     [,47]    [,48]     [,49]     [,50]     [,51]
#> [1,] -17.98025 -7.106741 -18.44314 13.08219 -7.799481 -17.48263 -2.806083
#>          [,52]      [,53]     [,54]     [,55]     [,56]    [,57]    [,58]
#> [1,] -11.92514 -0.9003706 -11.12341 -6.376393 -19.51989 9.287831 14.77298
#>        [,59]    [,60]     [,61]     [,62]    [,63]     [,64]    [,65]    [,66]
#> [1,] 3.13085 12.91601 -3.628517 -3.856418 12.99489 -6.540129 36.18738 3.396378
#>        [,67]  [,68]    [,69]     [,70]     [,71]    [,72]     [,73]     [,74]
#> [1,] 26.5798 17.274 2.295904 -8.195013 -24.44596 7.610938 -13.75352 -7.190293
#>          [,75]     [,76]     [,77]    [,78]     [,79]    [,80]     [,81]
#> [1,] -9.867442 -19.14275 -14.11561 18.03381 -8.166413 1.261207 -1.377538
#>          [,82]     [,83]     [,84]     [,85]    [,86]     [,87]   [,88]
#> [1,] -5.633228 0.3177746 -17.61411 0.8839497 -13.4329 -18.23417 -3.5298
#>         [,89]    [,90]    [,91]     [,92]     [,93]     [,94]     [,95]
#> [1,] -12.1792 7.758229 14.65329 -1.302101 -1.237959 -8.945242 -25.55273
#>         [,96]     [,97]    [,98]     [,99]     [,100]
#> [1,] 23.72911 -1.568364 2.751244 -2.984612 -0.6996492
integratemvn(matrix(1, 1, 1), 100L, c(5), matrix(1))
#>           [,1]     [,2]      [,3]    [,4]      [,5]     [,6]     [,7]     [,8]
#> [1,] -11.45241 1.013421 -1.265801 11.1725 -3.606036 4.073659 3.572474 2.790863
#>            [,9]     [,10]    [,11]     [,12]    [,13]     [,14]    [,15]
#> [1,] -0.4383182 -3.452375 0.739492 -3.806859 2.101751 0.9432718 2.495353
#>          [,16]     [,17]      [,18]      [,19]    [,20]    [,21]    [,22]
#> [1,] -9.831716 -8.488808 0.09702812 -0.3725566 0.618709 5.160829 1.032588
#>          [,23]    [,24]    [,25]     [,26]     [,27]    [,28]     [,29]
#> [1,] -6.835829 4.754478 1.087488 0.2511438 -2.452874 5.082315 -3.476889
#>          [,30]     [,31]    [,32]   [,33]     [,34]     [,35]     [,36]
#> [1,] 0.2570943 -2.603321 3.971077 -3.5269 -7.835584 -1.911467 -1.573441
#>         [,37]     [,38]     [,39]    [,40]    [,41]     [,42]    [,43]    [,44]
#> [1,] 5.080939 -9.292405 -5.175859 6.327392 4.417676 0.6784436 6.042248 4.368244
#>           [,45]     [,46]    [,47]    [,48]    [,49]     [,50]    [,51]
#> [1,] -0.7794436 -2.461815 -2.09521 9.935664 6.247498 -3.857963 2.688709
#>          [,52]     [,53]   [,54]     [,55]     [,56]    [,57]     [,58]
#> [1,] -6.031097 -6.330331 0.69127 -2.385782 -7.058173 3.497101 -4.875895
#>          [,59]     [,60]     [,61]     [,62]     [,63]     [,64]     [,65]
#> [1,] -1.692528 -6.642737 -5.039049 -7.910146 -3.128176 -2.369678 -3.766124
#>          [,66]     [,67]     [,68]    [,69]       [,70]     [,71]     [,72]
#> [1,] -1.607826 -4.106053 -3.319551 3.266706 -0.05670013 -4.506303 -1.270637
#>          [,73]    [,74]     [,75]    [,76]     [,77]     [,78]    [,79]
#> [1,] -4.667373 1.083874 -2.433586 11.62792 -7.498701 -2.634439 4.454745
#>          [,80]     [,81]    [,82]     [,83]     [,84]     [,85]     [,86]
#> [1,] -9.792099 -2.419386 6.710544 0.2220229 -5.283856 -3.441449 -6.322194
#>         [,87]      [,88]     [,89]     [,90]     [,91]   [,92]    [,93]   [,94]
#> [1,] 1.264824 -0.7865944 -2.736168 -2.547448 -13.34492 2.21708 4.773973 6.33409
#>         [,95]      [,96]    [,97]    [,98]      [,99]    [,100]
#> [1,] 8.130323 -0.8056644 6.186646 2.154267 -0.7424349 0.7603197